% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation.R
\name{Animation}
\alias{Animation}
\title{Animation handle}
\description{
Returned by \code{\link[=animate]{animate()}}.
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{widget, property}}{What is animated.}

    \item{\code{from, to}}{Start and end values.}

    \item{\code{duration}}{Seconds.}

    \item{\code{easing}}{Easing name.}

    \item{\code{start}}{Clock time of the start.}

    \item{\code{active}}{Still running?}

    \item{\code{on_complete}}{\verb{function(widget, app)} or \code{NULL}.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-Animation-initialize}{\code{Animation$new()}}
    \item \href{#method-Animation-cancel}{\code{Animation$cancel()}}
    \item \href{#method-Animation-clone}{\code{Animation$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Animation-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-Animation-initialize}{}}}
\subsection{\code{Animation$new()}}{
  Create an animation (use \code{\link[=animate]{animate()}}).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Animation$new(widget, property, from, to, duration, easing, start, on_complete)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{widget, property, from, to, duration, easing, start, on_complete}}{See fields.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Animation-cancel"></a>}}
\if{latex}{\out{\hypertarget{method-Animation-cancel}{}}}
\subsection{\code{Animation$cancel()}}{
  Stop the animation where it is.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Animation$cancel()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Animation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Animation-clone}{}}}
\subsection{\code{Animation$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Animation$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
