% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-controls.R
\name{Checkbox}
\alias{Checkbox}
\title{Checkbox widget}
\description{
A labelled boolean toggle. See \code{\link[=checkbox]{checkbox()}}.
}
\section{Super class}{
\code{\link[termr:Widget]{Widget}} -> \code{Checkbox}
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{focusable}}{Checkboxes can be focused.}
  }
  \if{html}{\out{</div>}}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{value}}{\code{TRUE} or \code{FALSE} (reactive; sends \code{"checkbox.changed"}).}

    \item{\code{label}}{The label (reactive).}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-Checkbox-initialize}{\code{Checkbox$new()}}
    \item \href{#method-Checkbox-default_style}{\code{Checkbox$default_style()}}
    \item \href{#method-Checkbox-default_bindings}{\code{Checkbox$default_bindings()}}
    \item \href{#method-Checkbox-toggle}{\code{Checkbox$toggle()}}
    \item \href{#method-Checkbox-action_toggle}{\code{Checkbox$action_toggle()}}
    \item \href{#method-Checkbox-on_click}{\code{Checkbox$on_click()}}
    \item \href{#method-Checkbox-render}{\code{Checkbox$render()}}
    \item \href{#method-Checkbox-clone}{\code{Checkbox$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="add_class"><a href='../../termr/html/Widget.html#method-Widget-add_class'><code>Widget$add_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="ancestors"><a href='../../termr/html/Widget.html#method-Widget-ancestors'><code>Widget$ancestors()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="animate"><a href='../../termr/html/Widget.html#method-Widget-animate'><code>Widget$animate()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="arrange_children"><a href='../../termr/html/Widget.html#method-Widget-arrange_children'><code>Widget$arrange_children()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="bind"><a href='../../termr/html/Widget.html#method-Widget-bind'><code>Widget$bind()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="bind_reactive"><a href='../../termr/html/Widget.html#method-Widget-bind_reactive'><code>Widget$bind_reactive()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="bindings"><a href='../../termr/html/Widget.html#method-Widget-bindings'><code>Widget$bindings()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="blur"><a href='../../termr/html/Widget.html#method-Widget-blur'><code>Widget$blur()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="capture_mouse"><a href='../../termr/html/Widget.html#method-Widget-capture_mouse'><code>Widget$capture_mouse()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="child_clip"><a href='../../termr/html/Widget.html#method-Widget-child_clip'><code>Widget$child_clip()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="clear"><a href='../../termr/html/Widget.html#method-Widget-clear'><code>Widget$clear()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="computed_style"><a href='../../termr/html/Widget.html#method-Widget-computed_style'><code>Widget$computed_style()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="content_height"><a href='../../termr/html/Widget.html#method-Widget-content_height'><code>Widget$content_height()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="content_width"><a href='../../termr/html/Widget.html#method-Widget-content_width'><code>Widget$content_width()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="focus"><a href='../../termr/html/Widget.html#method-Widget-focus'><code>Widget$focus()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="format"><a href='../../termr/html/Widget.html#method-Widget-format'><code>Widget$format()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="get_state"><a href='../../termr/html/Widget.html#method-Widget-get_state'><code>Widget$get_state()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="has_class"><a href='../../termr/html/Widget.html#method-Widget-has_class'><code>Widget$has_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="invalidate"><a href='../../termr/html/Widget.html#method-Widget-invalidate'><code>Widget$invalidate()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="invalidate_paint"><a href='../../termr/html/Widget.html#method-Widget-invalidate_paint'><code>Widget$invalidate_paint()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="is_displayed"><a href='../../termr/html/Widget.html#method-Widget-is_displayed'><code>Widget$is_displayed()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="is_enabled"><a href='../../termr/html/Widget.html#method-Widget-is_enabled'><code>Widget$is_enabled()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="layout_descend"><a href='../../termr/html/Widget.html#method-Widget-layout_descend'><code>Widget$layout_descend()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="layout_differs"><a href='../../termr/html/Widget.html#method-Widget-layout_differs'><code>Widget$layout_differs()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="matches"><a href='../../termr/html/Widget.html#method-Widget-matches'><code>Widget$matches()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="mount"><a href='../../termr/html/Widget.html#method-Widget-mount'><code>Widget$mount()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="on"><a href='../../termr/html/Widget.html#method-Widget-on'><code>Widget$on()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="paint"><a href='../../termr/html/Widget.html#method-Widget-paint'><code>Widget$paint()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="post_message"><a href='../../termr/html/Widget.html#method-Widget-post_message'><code>Widget$post_message()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="print"><a href='../../termr/html/Widget.html#method-Widget-print'><code>Widget$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="pseudo_states"><a href='../../termr/html/Widget.html#method-Widget-pseudo_states'><code>Widget$pseudo_states()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="query"><a href='../../termr/html/Widget.html#method-Widget-query'><code>Widget$query()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="query_one"><a href='../../termr/html/Widget.html#method-Widget-query_one'><code>Widget$query_one()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="refresh"><a href='../../termr/html/Widget.html#method-Widget-refresh'><code>Widget$refresh()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="release_mouse"><a href='../../termr/html/Widget.html#method-Widget-release_mouse'><code>Widget$release_mouse()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="remove"><a href='../../termr/html/Widget.html#method-Widget-remove'><code>Widget$remove()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="remove_children"><a href='../../termr/html/Widget.html#method-Widget-remove_children'><code>Widget$remove_children()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="remove_class"><a href='../../termr/html/Widget.html#method-Widget-remove_class'><code>Widget$remove_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="render_lines"><a href='../../termr/html/Widget.html#method-Widget-render_lines'><code>Widget$render_lines()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="replace"><a href='../../termr/html/Widget.html#method-Widget-replace'><code>Widget$replace()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="run_action"><a href='../../termr/html/Widget.html#method-Widget-run_action'><code>Widget$run_action()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="run_process"><a href='../../termr/html/Widget.html#method-Widget-run_process'><code>Widget$run_process()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="run_worker"><a href='../../termr/html/Widget.html#method-Widget-run_worker'><code>Widget$run_worker()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set"><a href='../../termr/html/Widget.html#method-Widget-set'><code>Widget$set()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set_interval"><a href='../../termr/html/Widget.html#method-Widget-set_interval'><code>Widget$set_interval()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set_state"><a href='../../termr/html/Widget.html#method-Widget-set_state'><code>Widget$set_state()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set_timeout"><a href='../../termr/html/Widget.html#method-Widget-set_timeout'><code>Widget$set_timeout()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="toggle_class"><a href='../../termr/html/Widget.html#method-Widget-toggle_class'><code>Widget$toggle_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="walk"><a href='../../termr/html/Widget.html#method-Widget-walk'><code>Widget$walk()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Checkbox-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-Checkbox-initialize}{}}}
\subsection{\code{Checkbox$new()}}{
  Create a checkbox. See \code{\link[=checkbox]{checkbox()}}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Checkbox$new(
  label = "",
  value = FALSE,
  id = NULL,
  classes = NULL,
  style = NULL,
  disabled = FALSE
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{label}}{Label.}
      \item{\code{value}}{Initial state.}
      \item{\code{id, classes, style, disabled}}{See \link{Widget}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Checkbox-default_style"></a>}}
\if{latex}{\out{\hypertarget{method-Checkbox-default_style}{}}}
\subsection{\code{Checkbox$default_style()}}{
  The built-in style.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Checkbox$default_style()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Checkbox-default_bindings"></a>}}
\if{latex}{\out{\hypertarget{method-Checkbox-default_bindings}{}}}
\subsection{\code{Checkbox$default_bindings()}}{
  Space and Enter toggle.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Checkbox$default_bindings()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Checkbox-toggle"></a>}}
\if{latex}{\out{\hypertarget{method-Checkbox-toggle}{}}}
\subsection{\code{Checkbox$toggle()}}{
  Toggle the value (sends \code{"checkbox.changed"}).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Checkbox$toggle()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Checkbox-action_toggle"></a>}}
\if{latex}{\out{\hypertarget{method-Checkbox-action_toggle}{}}}
\subsection{\code{Checkbox$action_toggle()}}{
  Action used by the bindings.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Checkbox$action_toggle()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Checkbox-on_click"></a>}}
\if{latex}{\out{\hypertarget{method-Checkbox-on_click}{}}}
\subsection{\code{Checkbox$on_click()}}{
  Clicking toggles.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Checkbox$on_click(event)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{event}}{A \code{MouseEvent}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Checkbox-render"></a>}}
\if{latex}{\out{\hypertarget{method-Checkbox-render}{}}}
\subsection{\code{Checkbox$render()}}{
  The box and the label.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Checkbox$render()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Checkbox-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Checkbox-clone}{}}}
\subsection{\code{Checkbox$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Checkbox$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
