% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-datatable.R
\name{DataTable}
\alias{DataTable}
\title{DataTable widget}
\description{
A virtualised table for data frames or lazy table sources. See \code{\link[=data_table]{data_table()}}.
}
\section{Super class}{
\code{\link[termr:Widget]{Widget}} -> \code{DataTable}
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{paint_states}}{Cursor and scroll changes only repaint.}

    \item{\code{focusable}}{Tables can be focused.}

    \item{\code{cursor_type}}{\code{"row"}, \code{"cell"} or \code{"none"}.}

    \item{\code{show_header}}{Show the header row?}

    \item{\code{zebra}}{Stripe alternate rows?}

    \item{\code{row_style}}{\code{NULL} or \verb{function(row, data)} returning a style for
a data row (\code{row} is the original row number or source position).}

    \item{\code{cell_style}}{\code{NULL} or \verb{function(value, row, column)} returning a
style for a cell.}

    \item{\code{frozen_columns}}{Number of leading columns that stay visible
while the others scroll horizontally.}

    \item{\code{header_sort}}{Clicking a header sorts by that column (click again:
descending, then unsorted; Shift+click adds a column)?}

    \item{\code{editable}}{Allow editing cells (Enter / double click)?}

    \item{\code{on_edit}}{\code{NULL} or \verb{function(row, column, value)} called before a
cell edit is applied; return \code{NULL} / \code{TRUE} to accept or a message
to reject the edit.}
  }
  \if{html}{\out{</div>}}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{data}}{The data frame or source (assigning calls \code{set_data()}).}

    \item{\code{row_count}}{Number of rows in the current view.}

    \item{\code{column_names}}{Names of the data columns.}

    \item{\code{cursor_row}}{Cursor position in the view (0 = none).}

    \item{\code{cursor_column}}{Cursor column number (cell cursor; 0 = none).}

    \item{\code{offset_row, offset_column}}{Scroll offsets (rows / columns).}

    \item{\code{view}}{Row numbers or source view positions in display order.}

    \item{\code{sort_state}}{\code{NULL} or a data frame (\code{column}, \code{decreasing}) of
the current sort keys.}

    \item{\code{match_hit}}{\code{NULL} or the current search result (\code{position},
\code{row}, \code{column}).}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-DataTable-initialize}{\code{DataTable$new()}}
    \item \href{#method-DataTable-default_style}{\code{DataTable$default_style()}}
    \item \href{#method-DataTable-default_bindings}{\code{DataTable$default_bindings()}}
    \item \href{#method-DataTable-set_data}{\code{DataTable$set_data()}}
    \item \href{#method-DataTable-refresh}{\code{DataTable$refresh()}}
    \item \href{#method-DataTable-close}{\code{DataTable$close()}}
    \item \href{#method-DataTable-source_stats}{\code{DataTable$source_stats()}}
    \item \href{#method-DataTable-sort}{\code{DataTable$sort()}}
    \item \href{#method-DataTable-clear_sort}{\code{DataTable$clear_sort()}}
    \item \href{#method-DataTable-toggle_sort}{\code{DataTable$toggle_sort()}}
    \item \href{#method-DataTable-filter}{\code{DataTable$filter()}}
    \item \href{#method-DataTable-filter_columns}{\code{DataTable$filter_columns()}}
    \item \href{#method-DataTable-move_cursor}{\code{DataTable$move_cursor()}}
    \item \href{#method-DataTable-scroll_to_row}{\code{DataTable$scroll_to_row()}}
    \item \href{#method-DataTable-selected_row}{\code{DataTable$selected_row()}}
    \item \href{#method-DataTable-row_data}{\code{DataTable$row_data()}}
    \item \href{#method-DataTable-selected_data}{\code{DataTable$selected_data()}}
    \item \href{#method-DataTable-visible_data}{\code{DataTable$visible_data()}}
    \item \href{#method-DataTable-reorder_column}{\code{DataTable$reorder_column()}}
    \item \href{#method-DataTable-select_range}{\code{DataTable$select_range()}}
    \item \href{#method-DataTable-copy_selection}{\code{DataTable$copy_selection()}}
    \item \href{#method-DataTable-goto_row}{\code{DataTable$goto_row()}}
    \item \href{#method-DataTable-column_width}{\code{DataTable$column_width()}}
    \item \href{#method-DataTable-set_column_width}{\code{DataTable$set_column_width()}}
    \item \href{#method-DataTable-auto_size_column}{\code{DataTable$auto_size_column()}}
    \item \href{#method-DataTable-auto_size_all}{\code{DataTable$auto_size_all()}}
    \item \href{#method-DataTable-set_column_visible}{\code{DataTable$set_column_visible()}}
    \item \href{#method-DataTable-column_info}{\code{DataTable$column_info()}}
    \item \href{#method-DataTable-find}{\code{DataTable$find()}}
    \item \href{#method-DataTable-find_next}{\code{DataTable$find_next()}}
    \item \href{#method-DataTable-find_previous}{\code{DataTable$find_previous()}}
    \item \href{#method-DataTable-clear_find}{\code{DataTable$clear_find()}}
    \item \href{#method-DataTable-edit_cell}{\code{DataTable$edit_cell()}}
    \item \href{#method-DataTable-set_cell}{\code{DataTable$set_cell()}}
    \item \href{#method-DataTable-action_cursor_up}{\code{DataTable$action_cursor_up()}}
    \item \href{#method-DataTable-action_cursor_down}{\code{DataTable$action_cursor_down()}}
    \item \href{#method-DataTable-action_extend_selection_up}{\code{DataTable$action_extend_selection_up()}}
    \item \href{#method-DataTable-action_extend_selection_down}{\code{DataTable$action_extend_selection_down()}}
    \item \href{#method-DataTable-action_move_column_left}{\code{DataTable$action_move_column_left()}}
    \item \href{#method-DataTable-action_move_column_right}{\code{DataTable$action_move_column_right()}}
    \item \href{#method-DataTable-action_copy_selection}{\code{DataTable$action_copy_selection()}}
    \item \href{#method-DataTable-action_cursor_left}{\code{DataTable$action_cursor_left()}}
    \item \href{#method-DataTable-action_cursor_right}{\code{DataTable$action_cursor_right()}}
    \item \href{#method-DataTable-action_page_up}{\code{DataTable$action_page_up()}}
    \item \href{#method-DataTable-action_page_down}{\code{DataTable$action_page_down()}}
    \item \href{#method-DataTable-action_first_row}{\code{DataTable$action_first_row()}}
    \item \href{#method-DataTable-action_last_row}{\code{DataTable$action_last_row()}}
    \item \href{#method-DataTable-action_open_find}{\code{DataTable$action_open_find()}}
    \item \href{#method-DataTable-action_open_goto}{\code{DataTable$action_open_goto()}}
    \item \href{#method-DataTable-action_find_next}{\code{DataTable$action_find_next()}}
    \item \href{#method-DataTable-action_find_previous}{\code{DataTable$action_find_previous()}}
    \item \href{#method-DataTable-action_widen_column}{\code{DataTable$action_widen_column()}}
    \item \href{#method-DataTable-action_narrow_column}{\code{DataTable$action_narrow_column()}}
    \item \href{#method-DataTable-action_sort_column}{\code{DataTable$action_sort_column()}}
    \item \href{#method-DataTable-action_activate}{\code{DataTable$action_activate()}}
    \item \href{#method-DataTable-on_mouse_down}{\code{DataTable$on_mouse_down()}}
    \item \href{#method-DataTable-on_drag_move}{\code{DataTable$on_drag_move()}}
    \item \href{#method-DataTable-on_drag_end}{\code{DataTable$on_drag_end()}}
    \item \href{#method-DataTable-on_key}{\code{DataTable$on_key()}}
    \item \href{#method-DataTable-on_paste}{\code{DataTable$on_paste()}}
    \item \href{#method-DataTable-on_mouse_scroll}{\code{DataTable$on_mouse_scroll()}}
    \item \href{#method-DataTable-content_width}{\code{DataTable$content_width()}}
    \item \href{#method-DataTable-content_height}{\code{DataTable$content_height()}}
    \item \href{#method-DataTable-paint}{\code{DataTable$paint()}}
    \item \href{#method-DataTable-clone}{\code{DataTable$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="add_class"><a href='../../termr/html/Widget.html#method-Widget-add_class'><code>Widget$add_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="ancestors"><a href='../../termr/html/Widget.html#method-Widget-ancestors'><code>Widget$ancestors()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="animate"><a href='../../termr/html/Widget.html#method-Widget-animate'><code>Widget$animate()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="arrange_children"><a href='../../termr/html/Widget.html#method-Widget-arrange_children'><code>Widget$arrange_children()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="bind"><a href='../../termr/html/Widget.html#method-Widget-bind'><code>Widget$bind()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="bind_reactive"><a href='../../termr/html/Widget.html#method-Widget-bind_reactive'><code>Widget$bind_reactive()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="bindings"><a href='../../termr/html/Widget.html#method-Widget-bindings'><code>Widget$bindings()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="blur"><a href='../../termr/html/Widget.html#method-Widget-blur'><code>Widget$blur()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="capture_mouse"><a href='../../termr/html/Widget.html#method-Widget-capture_mouse'><code>Widget$capture_mouse()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="child_clip"><a href='../../termr/html/Widget.html#method-Widget-child_clip'><code>Widget$child_clip()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="clear"><a href='../../termr/html/Widget.html#method-Widget-clear'><code>Widget$clear()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="computed_style"><a href='../../termr/html/Widget.html#method-Widget-computed_style'><code>Widget$computed_style()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="focus"><a href='../../termr/html/Widget.html#method-Widget-focus'><code>Widget$focus()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="format"><a href='../../termr/html/Widget.html#method-Widget-format'><code>Widget$format()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="get_state"><a href='../../termr/html/Widget.html#method-Widget-get_state'><code>Widget$get_state()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="has_class"><a href='../../termr/html/Widget.html#method-Widget-has_class'><code>Widget$has_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="invalidate"><a href='../../termr/html/Widget.html#method-Widget-invalidate'><code>Widget$invalidate()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="invalidate_paint"><a href='../../termr/html/Widget.html#method-Widget-invalidate_paint'><code>Widget$invalidate_paint()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="is_displayed"><a href='../../termr/html/Widget.html#method-Widget-is_displayed'><code>Widget$is_displayed()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="is_enabled"><a href='../../termr/html/Widget.html#method-Widget-is_enabled'><code>Widget$is_enabled()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="layout_descend"><a href='../../termr/html/Widget.html#method-Widget-layout_descend'><code>Widget$layout_descend()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="layout_differs"><a href='../../termr/html/Widget.html#method-Widget-layout_differs'><code>Widget$layout_differs()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="matches"><a href='../../termr/html/Widget.html#method-Widget-matches'><code>Widget$matches()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="mount"><a href='../../termr/html/Widget.html#method-Widget-mount'><code>Widget$mount()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="on"><a href='../../termr/html/Widget.html#method-Widget-on'><code>Widget$on()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="post_message"><a href='../../termr/html/Widget.html#method-Widget-post_message'><code>Widget$post_message()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="print"><a href='../../termr/html/Widget.html#method-Widget-print'><code>Widget$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="pseudo_states"><a href='../../termr/html/Widget.html#method-Widget-pseudo_states'><code>Widget$pseudo_states()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="query"><a href='../../termr/html/Widget.html#method-Widget-query'><code>Widget$query()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="query_one"><a href='../../termr/html/Widget.html#method-Widget-query_one'><code>Widget$query_one()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="release_mouse"><a href='../../termr/html/Widget.html#method-Widget-release_mouse'><code>Widget$release_mouse()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="remove"><a href='../../termr/html/Widget.html#method-Widget-remove'><code>Widget$remove()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="remove_children"><a href='../../termr/html/Widget.html#method-Widget-remove_children'><code>Widget$remove_children()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="remove_class"><a href='../../termr/html/Widget.html#method-Widget-remove_class'><code>Widget$remove_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="render"><a href='../../termr/html/Widget.html#method-Widget-render'><code>Widget$render()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="render_lines"><a href='../../termr/html/Widget.html#method-Widget-render_lines'><code>Widget$render_lines()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="replace"><a href='../../termr/html/Widget.html#method-Widget-replace'><code>Widget$replace()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="run_action"><a href='../../termr/html/Widget.html#method-Widget-run_action'><code>Widget$run_action()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="run_process"><a href='../../termr/html/Widget.html#method-Widget-run_process'><code>Widget$run_process()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="run_worker"><a href='../../termr/html/Widget.html#method-Widget-run_worker'><code>Widget$run_worker()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set"><a href='../../termr/html/Widget.html#method-Widget-set'><code>Widget$set()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set_interval"><a href='../../termr/html/Widget.html#method-Widget-set_interval'><code>Widget$set_interval()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set_state"><a href='../../termr/html/Widget.html#method-Widget-set_state'><code>Widget$set_state()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set_timeout"><a href='../../termr/html/Widget.html#method-Widget-set_timeout'><code>Widget$set_timeout()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="toggle_class"><a href='../../termr/html/Widget.html#method-Widget-toggle_class'><code>Widget$toggle_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="walk"><a href='../../termr/html/Widget.html#method-Widget-walk'><code>Widget$walk()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-initialize}{}}}
\subsection{\code{DataTable$new()}}{
  Create a table. See \code{\link[=data_table]{data_table()}}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$new(
  data,
  cursor = "row",
  columns = NULL,
  formatters = NULL,
  row_names = NA,
  header = TRUE,
  zebra = FALSE,
  row_style = NULL,
  cell_style = NULL,
  max_column_width = 40L,
  frozen_columns = 0L,
  header_sort = FALSE,
  editable = FALSE,
  on_edit = NULL,
  id = NULL,
  classes = NULL,
  style = NULL,
  disabled = FALSE
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{data}}{A data frame, matrix, or \code{\link[=table_source]{table_source()}}.}
      \item{\code{cursor}}{Cursor type.}
      \item{\code{columns}}{Named list of \code{\link[=column]{column()}} options.}
      \item{\code{formatters}}{Named list of formatter functions.}
      \item{\code{row_names}}{Show row names? \code{NA}: only when they are not 1..n.}
      \item{\code{header}}{Show the header?}
      \item{\code{zebra}}{Stripe rows?}
      \item{\code{row_style, cell_style}}{Style hooks.}
      \item{\code{max_column_width}}{Upper limit for estimated widths.}
      \item{\code{frozen_columns, header_sort, editable, on_edit}}{See the fields.}
      \item{\code{id, classes, style, disabled}}{See \link{Widget}.}
      \item{\code{filters}}{Named column filters for \code{filter_columns()}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-default_style"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-default_style}{}}}
\subsection{\code{DataTable$default_style()}}{
  The built-in style.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$default_style()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-default_bindings"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-default_bindings}{}}}
\subsection{\code{DataTable$default_bindings()}}{
  Navigation keys and Enter.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$default_bindings()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-set_data"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-set_data}{}}}
\subsection{\code{DataTable$set_data()}}{
  Replace the data. Resets the view, cursor and scroll.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$set_data(data)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{data}}{A data frame, matrix, or \code{\link[=table_source]{table_source()}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-refresh"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-refresh}{}}}
\subsection{\code{DataTable$refresh()}}{
  Re-read the source row count and clear cached rows.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$refresh()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-close"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-close}{}}}
\subsection{\code{DataTable$close()}}{
  Explicitly close this table's source resource.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$close()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-source_stats"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-source_stats}{}}}
\subsection{\code{DataTable$source_stats()}}{
  Read structural fetch/cache counters for diagnostics.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$source_stats()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-sort"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-sort}{}}}
\subsection{\code{DataTable$sort()}}{
  Sort the view by one or more columns (stable; \code{NA}s
last). Sorting always starts from the unsorted view (after
\code{filter()}), so \code{sort(c("cyl", "mpg"))} sorts by cyl, then mpg.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$sort(by, decreasing = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{by}}{Column names or numbers.}
      \item{\code{decreasing}}{Sort in decreasing order? One value, or one per
column.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-clear_sort"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-clear_sort}{}}}
\subsection{\code{DataTable$clear_sort()}}{
  Remove the sort order (back to the order before sorting).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$clear_sort()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-toggle_sort"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-toggle_sort}{}}}
\subsection{\code{DataTable$toggle_sort()}}{
  Header-click behaviour: sort by a column ascending, then
descending, then unsorted.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$toggle_sort(column, add = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{column}}{Column name or number.}
      \item{\code{add}}{Add the column to the existing sort (secondary key)?}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-filter"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-filter}{}}}
\subsection{\code{DataTable$filter()}}{
  Show a subset of rows (in the given order), or all rows.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$filter(rows = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{rows}}{Row numbers of the data, a logical vector over the rows,
or \code{NULL} for all rows.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-filter_columns"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-filter_columns}{}}}
\subsection{\code{DataTable$filter_columns()}}{
  Filter named columns; all predicates are combined with
AND and evaluated in row chunks. A predicate may be a
\code{table_filter()} or \code{function(values) logical}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$filter_columns(filters = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{filters}}{Named list by column; \code{NULL} clears column filters.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-move_cursor"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-move_cursor}{}}}
\subsection{\code{DataTable$move_cursor()}}{
  Move the cursor (and scroll so it is visible).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$move_cursor(row = NULL, column = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{row}}{Position in the view.}
      \item{\code{column}}{Column name or number (cell cursor).}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-scroll_to_row"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-scroll_to_row}{}}}
\subsection{\code{DataTable$scroll_to_row()}}{
  Scroll so that a view position is the first visible row.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$scroll_to_row(row)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{row}}{Position in the view.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-selected_row"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-selected_row}{}}}
\subsection{\code{DataTable$selected_row()}}{
  The data-frame row number or source row key under the cursor (or \code{NA}).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$selected_row()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-row_data"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-row_data}{}}}
\subsection{\code{DataTable$row_data()}}{
  One data row as a named list.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$row_data(row)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{row}}{Original row number.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-selected_data"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-selected_data}{}}}
\subsection{\code{DataTable$selected_data()}}{
  Data for selected view rows, in view order. A single
cursor row is returned when no range is selected.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$selected_data()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-visible_data"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-visible_data}{}}}
\subsection{\code{DataTable$visible_data()}}{
  Data in the current viewport, in display order.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$visible_data()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-reorder_column"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-reorder_column}{}}}
\subsection{\code{DataTable$reorder_column()}}{
  Move a column in the display order.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$reorder_column(column, to)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{column}}{Column name or number.}
      \item{\code{to}}{New 1-based position in the complete column order.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-select_range"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-select_range}{}}}
\subsection{\code{DataTable$select_range()}}{
  Select a range of rows by view position.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$select_range(from, to = from)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{from, to}}{1-based positions in the view.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-copy_selection"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-copy_selection}{}}}
\subsection{\code{DataTable$copy_selection()}}{
  Copy selected rows as tab-separated values.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$copy_selection(headers = TRUE, system = TRUE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{headers}}{Include column names?}
      \item{\code{system}}{Also write to the system clipboard when supported?}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-goto_row"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-goto_row}{}}}
\subsection{\code{DataTable$goto_row()}}{
  Jump to a row of the view (and scroll to it).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$goto_row(row)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{row}}{Position in the view (1-based).}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-column_width"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-column_width}{}}}
\subsection{\code{DataTable$column_width()}}{
  Width of a column in cells.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$column_width(column)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{column}}{Column name or number.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-set_column_width"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-set_column_width}{}}}
\subsection{\code{DataTable$set_column_width()}}{
  Set a column's width (at least 1 cell). Sends
\code{"datatable.column_resized"}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$set_column_width(column, width)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{column}}{Column name or number.}
      \item{\code{width}}{Width in cells.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-auto_size_column"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-auto_size_column}{}}}
\subsection{\code{DataTable$auto_size_column()}}{
  Fit a column to its content. Rows are sampled, never all
scanned unless asked for, so this is fast for huge tables.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$auto_size_column(column, rows = c("sample", "visible", "all"))}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{column}}{Column name or number.}
      \item{\code{rows}}{\code{"sample"} (the visible rows plus the first and last 500;
default), \code{"visible"} (visible rows only) or \code{"all"} (every row of
the view; slow for millions of rows).}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-auto_size_all"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-auto_size_all}{}}}
\subsection{\code{DataTable$auto_size_all()}}{
  Fit every visible column to its content.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$auto_size_all(rows = c("sample", "visible", "all"))}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{rows}}{See \code{auto_size_column()}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-set_column_visible"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-set_column_visible}{}}}
\subsection{\code{DataTable$set_column_visible()}}{
  Show or hide a column.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$set_column_visible(column, visible = TRUE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{column}}{Column name or number.}
      \item{\code{visible}}{\code{TRUE} to show, \code{FALSE} to hide.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-column_info"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-column_info}{}}}
\subsection{\code{DataTable$column_info()}}{
  Describe the columns: name, label, width, visible, sortable.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$column_info()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-find"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-find}{}}}
\subsection{\code{DataTable$find()}}{
  Search the formatted cell text and move the cursor to
the first match at or after the cursor (reading order, wrapping
around). Large tables are scanned in chunks; the data is never
converted to one big string matrix.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$find(query, column = NULL, case_sensitive = FALSE, regex = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{query}}{Text or regular expression (PCRE).}
      \item{\code{column}}{Search only this column (name or number); \code{NULL}: all
visible columns.}
      \item{\code{case_sensitive}}{Match case?}
      \item{\code{regex}}{Is \code{query} a regular expression?}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \code{TRUE} if there is a match (invisibly).
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-find_next"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-find_next}{}}}
\subsection{\code{DataTable$find_next()}}{
  Go to the next match.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$find_next()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-find_previous"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-find_previous}{}}}
\subsection{\code{DataTable$find_previous()}}{
  Go to the previous match.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$find_previous()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-clear_find"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-clear_find}{}}}
\subsection{\code{DataTable$clear_find()}}{
  Forget the search and close the find bar.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$clear_find()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-edit_cell"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-edit_cell}{}}}
\subsection{\code{DataTable$edit_cell()}}{
  Edit a cell in a small dialog (needs \code{editable = TRUE}).
The default is the cursor cell.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$edit_cell(position = NULL, column = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{position}}{Position in the view (default: the cursor row).}
      \item{\code{column}}{Column name or number (default: the cursor / active column).}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-set_cell"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-set_cell}{}}}
\subsection{\code{DataTable$set_cell()}}{
  Change one cell of the table's own copy of the data
(the data frame you passed in is never modified; read the result
with \verb{$data}). Runs \code{on_edit}, then sends \code{"datatable.cell_changed"}
with \code{row}, \code{column}, \code{old} and \code{value}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$set_cell(row, column, value)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{row}}{Original row number.}
      \item{\code{column}}{Column name or number.}
      \item{\code{value}}{New value (must fit the column type).}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-action_cursor_up"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-action_cursor_up}{}}}
\subsection{\code{DataTable$action_cursor_up()}}{
  Actions used by the bindings.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$action_cursor_up()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-action_cursor_down"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-action_cursor_down}{}}}
\subsection{\code{DataTable$action_cursor_down()}}{
  Move down.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$action_cursor_down()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-action_extend_selection_up"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-action_extend_selection_up}{}}}
\subsection{\code{DataTable$action_extend_selection_up()}}{
  Extend selection upward.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$action_extend_selection_up()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-action_extend_selection_down"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-action_extend_selection_down}{}}}
\subsection{\code{DataTable$action_extend_selection_down()}}{
  Extend selection downward.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$action_extend_selection_down()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-action_move_column_left"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-action_move_column_left}{}}}
\subsection{\code{DataTable$action_move_column_left()}}{
  Move the active column left in display order.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$action_move_column_left()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-action_move_column_right"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-action_move_column_right}{}}}
\subsection{\code{DataTable$action_move_column_right()}}{
  Move the active column right in display order.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$action_move_column_right()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-action_copy_selection"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-action_copy_selection}{}}}
\subsection{\code{DataTable$action_copy_selection()}}{
  Copy selected rows.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$action_copy_selection()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-action_cursor_left"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-action_cursor_left}{}}}
\subsection{\code{DataTable$action_cursor_left()}}{
  Move left (cell cursor) or scroll left.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$action_cursor_left()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-action_cursor_right"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-action_cursor_right}{}}}
\subsection{\code{DataTable$action_cursor_right()}}{
  Move right (cell cursor) or scroll right.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$action_cursor_right()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-action_page_up"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-action_page_up}{}}}
\subsection{\code{DataTable$action_page_up()}}{
  Move one page up.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$action_page_up()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-action_page_down"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-action_page_down}{}}}
\subsection{\code{DataTable$action_page_down()}}{
  Move one page down.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$action_page_down()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-action_first_row"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-action_first_row}{}}}
\subsection{\code{DataTable$action_first_row()}}{
  Go to the first row.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$action_first_row()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-action_last_row"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-action_last_row}{}}}
\subsection{\code{DataTable$action_last_row()}}{
  Go to the last row.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$action_last_row()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-action_open_find"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-action_open_find}{}}}
\subsection{\code{DataTable$action_open_find()}}{
  Open the find bar (Ctrl+F).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$action_open_find()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-action_open_goto"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-action_open_goto}{}}}
\subsection{\code{DataTable$action_open_goto()}}{
  Open the "go to row" bar (Ctrl+G).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$action_open_goto()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-action_find_next"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-action_find_next}{}}}
\subsection{\code{DataTable$action_find_next()}}{
  Go to the next match (F3).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$action_find_next()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-action_find_previous"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-action_find_previous}{}}}
\subsection{\code{DataTable$action_find_previous()}}{
  Go to the previous match (Shift+F3).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$action_find_previous()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-action_widen_column"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-action_widen_column}{}}}
\subsection{\code{DataTable$action_widen_column()}}{
  Make the active column one cell wider (Alt+Right).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$action_widen_column()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-action_narrow_column"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-action_narrow_column}{}}}
\subsection{\code{DataTable$action_narrow_column()}}{
  Make the active column one cell narrower (Alt+Left).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$action_narrow_column()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-action_sort_column"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-action_sort_column}{}}}
\subsection{\code{DataTable$action_sort_column()}}{
  Sort by the active column (S): ascending, descending, off.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$action_sort_column()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-action_activate"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-action_activate}{}}}
\subsection{\code{DataTable$action_activate()}}{
  Send \code{"datatable.row_activated"} for the cursor row
(or edit the cell when the table is editable).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$action_activate()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-on_mouse_down"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-on_mouse_down}{}}}
\subsection{\code{DataTable$on_mouse_down()}}{
  Mouse: clicking a row moves the cursor, clicking the
header sends \code{"datatable.header_selected"} (and sorts when
\code{header_sort} is on); dragging a column separator in the header
resizes the column; a double click edits a cell of an editable table.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$on_mouse_down(event)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{event}}{A \code{MouseEvent}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-on_drag_move"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-on_drag_move}{}}}
\subsection{\code{DataTable$on_drag_move()}}{
  Resizing a column by dragging its header separator.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$on_drag_move(event)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{event}}{A \code{MouseEvent}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-on_drag_end"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-on_drag_end}{}}}
\subsection{\code{DataTable$on_drag_end()}}{
  Ends a column resize.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$on_drag_end(event)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{event}}{A \code{MouseEvent}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-on_key"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-on_key}{}}}
\subsection{\code{DataTable$on_key()}}{
  Keys for the find / go-to bar while it is open.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$on_key(event)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{event}}{A \code{KeyEvent}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-on_paste"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-on_paste}{}}}
\subsection{\code{DataTable$on_paste()}}{
  Pasted text goes into the find / go-to bar when open.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$on_paste(event)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{event}}{A \code{PasteEvent}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-on_mouse_scroll"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-on_mouse_scroll}{}}}
\subsection{\code{DataTable$on_mouse_scroll()}}{
  The mouse wheel scrolls three rows.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$on_mouse_scroll(event)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{event}}{A \code{MouseEvent}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-content_width"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-content_width}{}}}
\subsection{\code{DataTable$content_width()}}{
  Natural width: all columns.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$content_width()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-content_height"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-content_height}{}}}
\subsection{\code{DataTable$content_height()}}{
  Natural height: header plus rows (at most 1000).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$content_height(width)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{width}}{Content width.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-paint"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-paint}{}}}
\subsection{\code{DataTable$paint()}}{
  Draw the visible part of the table.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$paint(buffer, area, st)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{buffer}}{A \link{ScreenBuffer}.}
      \item{\code{area}}{Visible part of the region.}
      \item{\code{st}}{Computed style.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataTable-clone}{}}}
\subsection{\code{DataTable$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataTable$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
