% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-input.R
\name{Input}
\alias{Input}
\title{Input widget}
\description{
A single-line text input. See \code{\link[=input]{input()}}.
}
\section{Super class}{
\code{\link[termr:Widget]{Widget}} -> \code{Input}
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{paint_states}}{Cursor and selection changes only repaint.}

    \item{\code{focusable}}{Inputs can be focused.}

    \item{\code{placeholder}}{Text shown while the input is empty.}

    \item{\code{password}}{Show bullets instead of the characters?}

    \item{\code{max_length}}{Maximum number of characters, or \code{NULL}.}

    \item{\code{validate}}{\code{NULL} or \verb{function(value)} returning \code{NULL} (valid)
or an error message.}
  }
  \if{html}{\out{</div>}}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{value}}{The text (reactive). Assigning moves the cursor to the
end and sends \code{"input.changed"}.}

    \item{\code{cursor_position}}{Cursor position in characters (0 = start).}

    \item{\code{selection}}{The selected text (\code{""} if none).}

    \item{\code{valid}}{Does the value pass \code{validate}?}

    \item{\code{error}}{The validation message, or \code{NULL}.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-Input-initialize}{\code{Input$new()}}
    \item \href{#method-Input-default_style}{\code{Input$default_style()}}
    \item \href{#method-Input-pseudo_states}{\code{Input$pseudo_states()}}
    \item \href{#method-Input-default_bindings}{\code{Input$default_bindings()}}
    \item \href{#method-Input-on_key}{\code{Input$on_key()}}
    \item \href{#method-Input-on_mouse_down}{\code{Input$on_mouse_down()}}
    \item \href{#method-Input-on_mouse_move}{\code{Input$on_mouse_move()}}
    \item \href{#method-Input-on_paste}{\code{Input$on_paste()}}
    \item \href{#method-Input-insert}{\code{Input$insert()}}
    \item \href{#method-Input-clear}{\code{Input$clear()}}
    \item \href{#method-Input-select_range}{\code{Input$select_range()}}
    \item \href{#method-Input-copy}{\code{Input$copy()}}
    \item \href{#method-Input-action_cursor_left}{\code{Input$action_cursor_left()}}
    \item \href{#method-Input-action_cursor_right}{\code{Input$action_cursor_right()}}
    \item \href{#method-Input-action_cursor_home}{\code{Input$action_cursor_home()}}
    \item \href{#method-Input-action_cursor_end}{\code{Input$action_cursor_end()}}
    \item \href{#method-Input-action_word_left}{\code{Input$action_word_left()}}
    \item \href{#method-Input-action_word_right}{\code{Input$action_word_right()}}
    \item \href{#method-Input-action_select_left}{\code{Input$action_select_left()}}
    \item \href{#method-Input-action_select_right}{\code{Input$action_select_right()}}
    \item \href{#method-Input-action_select_home}{\code{Input$action_select_home()}}
    \item \href{#method-Input-action_select_end}{\code{Input$action_select_end()}}
    \item \href{#method-Input-action_select_word_left}{\code{Input$action_select_word_left()}}
    \item \href{#method-Input-action_select_word_right}{\code{Input$action_select_word_right()}}
    \item \href{#method-Input-action_select_all}{\code{Input$action_select_all()}}
    \item \href{#method-Input-action_delete_left}{\code{Input$action_delete_left()}}
    \item \href{#method-Input-action_delete_right}{\code{Input$action_delete_right()}}
    \item \href{#method-Input-action_delete_word_left}{\code{Input$action_delete_word_left()}}
    \item \href{#method-Input-action_delete_word_right}{\code{Input$action_delete_word_right()}}
    \item \href{#method-Input-action_delete_to_start}{\code{Input$action_delete_to_start()}}
    \item \href{#method-Input-action_delete_to_end}{\code{Input$action_delete_to_end()}}
    \item \href{#method-Input-action_paste}{\code{Input$action_paste()}}
    \item \href{#method-Input-action_submit}{\code{Input$action_submit()}}
    \item \href{#method-Input-content_width}{\code{Input$content_width()}}
    \item \href{#method-Input-render_lines}{\code{Input$render_lines()}}
    \item \href{#method-Input-clone}{\code{Input$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="add_class"><a href='../../termr/html/Widget.html#method-Widget-add_class'><code>Widget$add_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="ancestors"><a href='../../termr/html/Widget.html#method-Widget-ancestors'><code>Widget$ancestors()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="animate"><a href='../../termr/html/Widget.html#method-Widget-animate'><code>Widget$animate()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="arrange_children"><a href='../../termr/html/Widget.html#method-Widget-arrange_children'><code>Widget$arrange_children()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="bind"><a href='../../termr/html/Widget.html#method-Widget-bind'><code>Widget$bind()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="bind_reactive"><a href='../../termr/html/Widget.html#method-Widget-bind_reactive'><code>Widget$bind_reactive()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="bindings"><a href='../../termr/html/Widget.html#method-Widget-bindings'><code>Widget$bindings()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="blur"><a href='../../termr/html/Widget.html#method-Widget-blur'><code>Widget$blur()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="capture_mouse"><a href='../../termr/html/Widget.html#method-Widget-capture_mouse'><code>Widget$capture_mouse()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="child_clip"><a href='../../termr/html/Widget.html#method-Widget-child_clip'><code>Widget$child_clip()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="computed_style"><a href='../../termr/html/Widget.html#method-Widget-computed_style'><code>Widget$computed_style()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="content_height"><a href='../../termr/html/Widget.html#method-Widget-content_height'><code>Widget$content_height()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="focus"><a href='../../termr/html/Widget.html#method-Widget-focus'><code>Widget$focus()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="format"><a href='../../termr/html/Widget.html#method-Widget-format'><code>Widget$format()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="get_state"><a href='../../termr/html/Widget.html#method-Widget-get_state'><code>Widget$get_state()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="has_class"><a href='../../termr/html/Widget.html#method-Widget-has_class'><code>Widget$has_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="invalidate"><a href='../../termr/html/Widget.html#method-Widget-invalidate'><code>Widget$invalidate()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="invalidate_paint"><a href='../../termr/html/Widget.html#method-Widget-invalidate_paint'><code>Widget$invalidate_paint()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="is_displayed"><a href='../../termr/html/Widget.html#method-Widget-is_displayed'><code>Widget$is_displayed()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="is_enabled"><a href='../../termr/html/Widget.html#method-Widget-is_enabled'><code>Widget$is_enabled()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="layout_descend"><a href='../../termr/html/Widget.html#method-Widget-layout_descend'><code>Widget$layout_descend()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="layout_differs"><a href='../../termr/html/Widget.html#method-Widget-layout_differs'><code>Widget$layout_differs()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="matches"><a href='../../termr/html/Widget.html#method-Widget-matches'><code>Widget$matches()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="mount"><a href='../../termr/html/Widget.html#method-Widget-mount'><code>Widget$mount()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="on"><a href='../../termr/html/Widget.html#method-Widget-on'><code>Widget$on()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="paint"><a href='../../termr/html/Widget.html#method-Widget-paint'><code>Widget$paint()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="post_message"><a href='../../termr/html/Widget.html#method-Widget-post_message'><code>Widget$post_message()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="print"><a href='../../termr/html/Widget.html#method-Widget-print'><code>Widget$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="query"><a href='../../termr/html/Widget.html#method-Widget-query'><code>Widget$query()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="query_one"><a href='../../termr/html/Widget.html#method-Widget-query_one'><code>Widget$query_one()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="refresh"><a href='../../termr/html/Widget.html#method-Widget-refresh'><code>Widget$refresh()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="release_mouse"><a href='../../termr/html/Widget.html#method-Widget-release_mouse'><code>Widget$release_mouse()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="remove"><a href='../../termr/html/Widget.html#method-Widget-remove'><code>Widget$remove()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="remove_children"><a href='../../termr/html/Widget.html#method-Widget-remove_children'><code>Widget$remove_children()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="remove_class"><a href='../../termr/html/Widget.html#method-Widget-remove_class'><code>Widget$remove_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="render"><a href='../../termr/html/Widget.html#method-Widget-render'><code>Widget$render()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="replace"><a href='../../termr/html/Widget.html#method-Widget-replace'><code>Widget$replace()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="run_action"><a href='../../termr/html/Widget.html#method-Widget-run_action'><code>Widget$run_action()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="run_process"><a href='../../termr/html/Widget.html#method-Widget-run_process'><code>Widget$run_process()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="run_worker"><a href='../../termr/html/Widget.html#method-Widget-run_worker'><code>Widget$run_worker()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set"><a href='../../termr/html/Widget.html#method-Widget-set'><code>Widget$set()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set_interval"><a href='../../termr/html/Widget.html#method-Widget-set_interval'><code>Widget$set_interval()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set_state"><a href='../../termr/html/Widget.html#method-Widget-set_state'><code>Widget$set_state()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set_timeout"><a href='../../termr/html/Widget.html#method-Widget-set_timeout'><code>Widget$set_timeout()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="toggle_class"><a href='../../termr/html/Widget.html#method-Widget-toggle_class'><code>Widget$toggle_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="walk"><a href='../../termr/html/Widget.html#method-Widget-walk'><code>Widget$walk()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-Input-initialize}{}}}
\subsection{\code{Input$new()}}{
  Create an input.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$new(
  value = "",
  placeholder = "",
  id = NULL,
  classes = NULL,
  style = NULL,
  disabled = FALSE,
  password = FALSE,
  max_length = NULL,
  validate = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{value}}{Initial text.}
      \item{\code{placeholder}}{Placeholder text.}
      \item{\code{id, classes, style, disabled}}{See \link{Widget}.}
      \item{\code{password}}{Mask the characters?}
      \item{\code{max_length}}{Maximum number of characters.}
      \item{\code{validate}}{Validation function.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-default_style"></a>}}
\if{latex}{\out{\hypertarget{method-Input-default_style}{}}}
\subsection{\code{Input$default_style()}}{
  The built-in style.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$default_style()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-pseudo_states"></a>}}
\if{latex}{\out{\hypertarget{method-Input-pseudo_states}{}}}
\subsection{\code{Input$pseudo_states()}}{
  Active states: focus, hover, disabled and invalid.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$pseudo_states()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-default_bindings"></a>}}
\if{latex}{\out{\hypertarget{method-Input-default_bindings}{}}}
\subsection{\code{Input$default_bindings()}}{
  Editing key bindings.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$default_bindings()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-on_key"></a>}}
\if{latex}{\out{\hypertarget{method-Input-on_key}{}}}
\subsection{\code{Input$on_key()}}{
  Printable keys replace the selection or are inserted at
the cursor. Ctrl+C / Ctrl+X copy / cut when text is selected (otherwise
Ctrl+C keeps its app meaning).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$on_key(event)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{event}}{A \code{KeyEvent}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-on_mouse_down"></a>}}
\if{latex}{\out{\hypertarget{method-Input-on_mouse_down}{}}}
\subsection{\code{Input$on_mouse_down()}}{
  Clicking places the cursor; shift+click and dragging
select.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$on_mouse_down(event)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{event}}{A \code{MouseEvent}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-on_mouse_move"></a>}}
\if{latex}{\out{\hypertarget{method-Input-on_mouse_move}{}}}
\subsection{\code{Input$on_mouse_move()}}{
  Dragging with the left button extends the selection.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$on_mouse_move(event)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{event}}{A \code{MouseEvent}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-on_paste"></a>}}
\if{latex}{\out{\hypertarget{method-Input-on_paste}{}}}
\subsection{\code{Input$on_paste()}}{
  Pasted text is inserted as one edit. An Input is a
single line, so line breaks in the paste become spaces.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$on_paste(event)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{event}}{A \code{PasteEvent}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-insert"></a>}}
\if{latex}{\out{\hypertarget{method-Input-insert}{}}}
\subsection{\code{Input$insert()}}{
  Insert text at the cursor (replacing the selection).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$insert(text)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{text}}{A string.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-clear"></a>}}
\if{latex}{\out{\hypertarget{method-Input-clear}{}}}
\subsection{\code{Input$clear()}}{
  Remove all text.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$clear()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-select_range"></a>}}
\if{latex}{\out{\hypertarget{method-Input-select_range}{}}}
\subsection{\code{Input$select_range()}}{
  Select a range of characters.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$select_range(start, end)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{start, end}}{Positions (0 = before the first character).}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-copy"></a>}}
\if{latex}{\out{\hypertarget{method-Input-copy}{}}}
\subsection{\code{Input$copy()}}{
  Copy the selection to the app clipboard (and, when
\code{system = TRUE} and supported, to the system clipboard via OSC 52).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$copy(system = TRUE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{system}}{Also write to the system clipboard?}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-action_cursor_left"></a>}}
\if{latex}{\out{\hypertarget{method-Input-action_cursor_left}{}}}
\subsection{\code{Input$action_cursor_left()}}{
  Move one character left.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$action_cursor_left()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-action_cursor_right"></a>}}
\if{latex}{\out{\hypertarget{method-Input-action_cursor_right}{}}}
\subsection{\code{Input$action_cursor_right()}}{
  Move one character right.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$action_cursor_right()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-action_cursor_home"></a>}}
\if{latex}{\out{\hypertarget{method-Input-action_cursor_home}{}}}
\subsection{\code{Input$action_cursor_home()}}{
  Move to the start.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$action_cursor_home()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-action_cursor_end"></a>}}
\if{latex}{\out{\hypertarget{method-Input-action_cursor_end}{}}}
\subsection{\code{Input$action_cursor_end()}}{
  Move to the end.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$action_cursor_end()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-action_word_left"></a>}}
\if{latex}{\out{\hypertarget{method-Input-action_word_left}{}}}
\subsection{\code{Input$action_word_left()}}{
  Move to the previous word.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$action_word_left()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-action_word_right"></a>}}
\if{latex}{\out{\hypertarget{method-Input-action_word_right}{}}}
\subsection{\code{Input$action_word_right()}}{
  Move to the next word.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$action_word_right()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-action_select_left"></a>}}
\if{latex}{\out{\hypertarget{method-Input-action_select_left}{}}}
\subsection{\code{Input$action_select_left()}}{
  Extend the selection left.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$action_select_left()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-action_select_right"></a>}}
\if{latex}{\out{\hypertarget{method-Input-action_select_right}{}}}
\subsection{\code{Input$action_select_right()}}{
  Extend the selection right.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$action_select_right()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-action_select_home"></a>}}
\if{latex}{\out{\hypertarget{method-Input-action_select_home}{}}}
\subsection{\code{Input$action_select_home()}}{
  Select to the start.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$action_select_home()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-action_select_end"></a>}}
\if{latex}{\out{\hypertarget{method-Input-action_select_end}{}}}
\subsection{\code{Input$action_select_end()}}{
  Select to the end.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$action_select_end()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-action_select_word_left"></a>}}
\if{latex}{\out{\hypertarget{method-Input-action_select_word_left}{}}}
\subsection{\code{Input$action_select_word_left()}}{
  Select to the previous word.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$action_select_word_left()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-action_select_word_right"></a>}}
\if{latex}{\out{\hypertarget{method-Input-action_select_word_right}{}}}
\subsection{\code{Input$action_select_word_right()}}{
  Select to the next word.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$action_select_word_right()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-action_select_all"></a>}}
\if{latex}{\out{\hypertarget{method-Input-action_select_all}{}}}
\subsection{\code{Input$action_select_all()}}{
  Select everything.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$action_select_all()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-action_delete_left"></a>}}
\if{latex}{\out{\hypertarget{method-Input-action_delete_left}{}}}
\subsection{\code{Input$action_delete_left()}}{
  Delete the selection or the character before the cursor.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$action_delete_left()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-action_delete_right"></a>}}
\if{latex}{\out{\hypertarget{method-Input-action_delete_right}{}}}
\subsection{\code{Input$action_delete_right()}}{
  Delete the selection or the character at the cursor.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$action_delete_right()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-action_delete_word_left"></a>}}
\if{latex}{\out{\hypertarget{method-Input-action_delete_word_left}{}}}
\subsection{\code{Input$action_delete_word_left()}}{
  Delete the word before the cursor.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$action_delete_word_left()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-action_delete_word_right"></a>}}
\if{latex}{\out{\hypertarget{method-Input-action_delete_word_right}{}}}
\subsection{\code{Input$action_delete_word_right()}}{
  Delete the word after the cursor.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$action_delete_word_right()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-action_delete_to_start"></a>}}
\if{latex}{\out{\hypertarget{method-Input-action_delete_to_start}{}}}
\subsection{\code{Input$action_delete_to_start()}}{
  Delete everything before the cursor.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$action_delete_to_start()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-action_delete_to_end"></a>}}
\if{latex}{\out{\hypertarget{method-Input-action_delete_to_end}{}}}
\subsection{\code{Input$action_delete_to_end()}}{
  Delete everything after the cursor.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$action_delete_to_end()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-action_paste"></a>}}
\if{latex}{\out{\hypertarget{method-Input-action_paste}{}}}
\subsection{\code{Input$action_paste()}}{
  Insert the app clipboard.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$action_paste()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-action_submit"></a>}}
\if{latex}{\out{\hypertarget{method-Input-action_submit}{}}}
\subsection{\code{Input$action_submit()}}{
  Send an \code{"input.submitted"} message (Enter).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$action_submit()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-content_width"></a>}}
\if{latex}{\out{\hypertarget{method-Input-content_width}{}}}
\subsection{\code{Input$content_width()}}{
  Natural content width.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$content_width()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-render_lines"></a>}}
\if{latex}{\out{\hypertarget{method-Input-render_lines}{}}}
\subsection{\code{Input$render_lines()}}{
  The visible part of the text, scrolled so the cursor
is visible, with the cursor drawn in reverse video and the
selection highlighted when focused.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$render_lines(width = NA_integer_)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{width}}{Content width.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Input-clone}{}}}
\subsection{\code{Input$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Input$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
