% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{key_event}
\alias{key_event}
\title{Create a key event}
\usage{
key_event(key)
}
\arguments{
\item{key}{Key name such as \code{"a"}, \code{"enter"}, \code{"ctrl+c"} or
\code{"shift+tab"}.}
}
\value{
A \code{KeyEvent}.
}
\description{
Mostly useful in tests: see \code{\link[=test_app]{test_app()}} for simulating key presses.
}
\examples{
key_event("ctrl+c")
}
