% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-display.R
\name{Metric}
\alias{Metric}
\title{Metric widget}
\description{
See \code{\link[=metric]{metric()}}.
}
\section{Super class}{
\code{\link[termr:Widget]{Widget}} -> \code{Metric}
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{formatter}}{\verb{function(value)} returning a string.}
  }
  \if{html}{\out{</div>}}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{value}}{The value (reactive).}

    \item{\code{delta}}{The change shown next to the value (reactive).}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-Metric-initialize}{\code{Metric$new()}}
    \item \href{#method-Metric-default_style}{\code{Metric$default_style()}}
    \item \href{#method-Metric-update}{\code{Metric$update()}}
    \item \href{#method-Metric-render}{\code{Metric$render()}}
    \item \href{#method-Metric-clone}{\code{Metric$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="add_class"><a href='../../termr/html/Widget.html#method-Widget-add_class'><code>Widget$add_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="ancestors"><a href='../../termr/html/Widget.html#method-Widget-ancestors'><code>Widget$ancestors()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="animate"><a href='../../termr/html/Widget.html#method-Widget-animate'><code>Widget$animate()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="arrange_children"><a href='../../termr/html/Widget.html#method-Widget-arrange_children'><code>Widget$arrange_children()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="bind"><a href='../../termr/html/Widget.html#method-Widget-bind'><code>Widget$bind()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="bind_reactive"><a href='../../termr/html/Widget.html#method-Widget-bind_reactive'><code>Widget$bind_reactive()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="bindings"><a href='../../termr/html/Widget.html#method-Widget-bindings'><code>Widget$bindings()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="blur"><a href='../../termr/html/Widget.html#method-Widget-blur'><code>Widget$blur()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="capture_mouse"><a href='../../termr/html/Widget.html#method-Widget-capture_mouse'><code>Widget$capture_mouse()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="child_clip"><a href='../../termr/html/Widget.html#method-Widget-child_clip'><code>Widget$child_clip()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="clear"><a href='../../termr/html/Widget.html#method-Widget-clear'><code>Widget$clear()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="computed_style"><a href='../../termr/html/Widget.html#method-Widget-computed_style'><code>Widget$computed_style()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="content_height"><a href='../../termr/html/Widget.html#method-Widget-content_height'><code>Widget$content_height()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="content_width"><a href='../../termr/html/Widget.html#method-Widget-content_width'><code>Widget$content_width()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="default_bindings"><a href='../../termr/html/Widget.html#method-Widget-default_bindings'><code>Widget$default_bindings()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="focus"><a href='../../termr/html/Widget.html#method-Widget-focus'><code>Widget$focus()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="format"><a href='../../termr/html/Widget.html#method-Widget-format'><code>Widget$format()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="get_state"><a href='../../termr/html/Widget.html#method-Widget-get_state'><code>Widget$get_state()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="has_class"><a href='../../termr/html/Widget.html#method-Widget-has_class'><code>Widget$has_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="invalidate"><a href='../../termr/html/Widget.html#method-Widget-invalidate'><code>Widget$invalidate()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="invalidate_paint"><a href='../../termr/html/Widget.html#method-Widget-invalidate_paint'><code>Widget$invalidate_paint()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="is_displayed"><a href='../../termr/html/Widget.html#method-Widget-is_displayed'><code>Widget$is_displayed()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="is_enabled"><a href='../../termr/html/Widget.html#method-Widget-is_enabled'><code>Widget$is_enabled()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="layout_descend"><a href='../../termr/html/Widget.html#method-Widget-layout_descend'><code>Widget$layout_descend()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="layout_differs"><a href='../../termr/html/Widget.html#method-Widget-layout_differs'><code>Widget$layout_differs()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="matches"><a href='../../termr/html/Widget.html#method-Widget-matches'><code>Widget$matches()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="mount"><a href='../../termr/html/Widget.html#method-Widget-mount'><code>Widget$mount()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="on"><a href='../../termr/html/Widget.html#method-Widget-on'><code>Widget$on()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="paint"><a href='../../termr/html/Widget.html#method-Widget-paint'><code>Widget$paint()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="post_message"><a href='../../termr/html/Widget.html#method-Widget-post_message'><code>Widget$post_message()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="print"><a href='../../termr/html/Widget.html#method-Widget-print'><code>Widget$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="pseudo_states"><a href='../../termr/html/Widget.html#method-Widget-pseudo_states'><code>Widget$pseudo_states()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="query"><a href='../../termr/html/Widget.html#method-Widget-query'><code>Widget$query()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="query_one"><a href='../../termr/html/Widget.html#method-Widget-query_one'><code>Widget$query_one()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="refresh"><a href='../../termr/html/Widget.html#method-Widget-refresh'><code>Widget$refresh()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="release_mouse"><a href='../../termr/html/Widget.html#method-Widget-release_mouse'><code>Widget$release_mouse()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="remove"><a href='../../termr/html/Widget.html#method-Widget-remove'><code>Widget$remove()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="remove_children"><a href='../../termr/html/Widget.html#method-Widget-remove_children'><code>Widget$remove_children()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="remove_class"><a href='../../termr/html/Widget.html#method-Widget-remove_class'><code>Widget$remove_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="render_lines"><a href='../../termr/html/Widget.html#method-Widget-render_lines'><code>Widget$render_lines()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="replace"><a href='../../termr/html/Widget.html#method-Widget-replace'><code>Widget$replace()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="run_action"><a href='../../termr/html/Widget.html#method-Widget-run_action'><code>Widget$run_action()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="run_process"><a href='../../termr/html/Widget.html#method-Widget-run_process'><code>Widget$run_process()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="run_worker"><a href='../../termr/html/Widget.html#method-Widget-run_worker'><code>Widget$run_worker()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set"><a href='../../termr/html/Widget.html#method-Widget-set'><code>Widget$set()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set_interval"><a href='../../termr/html/Widget.html#method-Widget-set_interval'><code>Widget$set_interval()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set_state"><a href='../../termr/html/Widget.html#method-Widget-set_state'><code>Widget$set_state()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set_timeout"><a href='../../termr/html/Widget.html#method-Widget-set_timeout'><code>Widget$set_timeout()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="toggle_class"><a href='../../termr/html/Widget.html#method-Widget-toggle_class'><code>Widget$toggle_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="walk"><a href='../../termr/html/Widget.html#method-Widget-walk'><code>Widget$walk()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metric-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-Metric-initialize}{}}}
\subsection{\code{Metric$new()}}{
  Create a metric. See \code{\link[=metric]{metric()}}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Metric$new(
  label,
  value = NA,
  delta = NULL,
  format = NULL,
  id = NULL,
  classes = NULL,
  style = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{label}}{Label.}
      \item{\code{value}}{Value.}
      \item{\code{delta}}{Change.}
      \item{\code{format}}{Formatter.}
      \item{\code{id, classes, style}}{See \link{Widget}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metric-default_style"></a>}}
\if{latex}{\out{\hypertarget{method-Metric-default_style}{}}}
\subsection{\code{Metric$default_style()}}{
  The built-in style.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Metric$default_style()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metric-update"></a>}}
\if{latex}{\out{\hypertarget{method-Metric-update}{}}}
\subsection{\code{Metric$update()}}{
  Set the value (and optionally the delta).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Metric$update(value, delta = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{value}}{New value.}
      \item{\code{delta}}{New delta (\code{NULL} keeps none).}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metric-render"></a>}}
\if{latex}{\out{\hypertarget{method-Metric-render}{}}}
\subsection{\code{Metric$render()}}{
  Label, value and delta lines.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Metric$render()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metric-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Metric-clone}{}}}
\subsection{\code{Metric$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Metric$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
