% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-process.R
\name{ProcessView}
\alias{ProcessView}
\title{ProcessView widget}
\description{
Runs a program and shows its output. See \code{\link[=process_view]{process_view()}}.
}
\section{Super classes}{
\code{\link[termr:Widget]{Widget}} -> \code{\link[termr:Vertical]{Vertical}} -> \code{ProcessView}
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{command, args, wd, env, timeout}}{The process to run.}

    \item{\code{autostart}}{Start when the widget is mounted?}

    \item{\code{worker}}{The \link{Worker} of the current run, or \code{NULL}.}
  }
  \if{html}{\out{</div>}}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{running}}{Is the process running?}

    \item{\code{output}}{The output lines so far.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-ProcessView-initialize}{\code{ProcessView$new()}}
    \item \href{#method-ProcessView-default_bindings}{\code{ProcessView$default_bindings()}}
    \item \href{#method-ProcessView-on_mount}{\code{ProcessView$on_mount()}}
    \item \href{#method-ProcessView-start}{\code{ProcessView$start()}}
    \item \href{#method-ProcessView-cancel}{\code{ProcessView$cancel()}}
    \item \href{#method-ProcessView-restart}{\code{ProcessView$restart()}}
    \item \href{#method-ProcessView-action_cancel_process}{\code{ProcessView$action_cancel_process()}}
    \item \href{#method-ProcessView-action_restart_process}{\code{ProcessView$action_restart_process()}}
    \item \href{#method-ProcessView-clone}{\code{ProcessView$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="add_class"><a href='../../termr/html/Widget.html#method-Widget-add_class'><code>Widget$add_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="ancestors"><a href='../../termr/html/Widget.html#method-Widget-ancestors'><code>Widget$ancestors()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="animate"><a href='../../termr/html/Widget.html#method-Widget-animate'><code>Widget$animate()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="arrange_children"><a href='../../termr/html/Widget.html#method-Widget-arrange_children'><code>Widget$arrange_children()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="bind"><a href='../../termr/html/Widget.html#method-Widget-bind'><code>Widget$bind()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="bind_reactive"><a href='../../termr/html/Widget.html#method-Widget-bind_reactive'><code>Widget$bind_reactive()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="bindings"><a href='../../termr/html/Widget.html#method-Widget-bindings'><code>Widget$bindings()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="blur"><a href='../../termr/html/Widget.html#method-Widget-blur'><code>Widget$blur()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="capture_mouse"><a href='../../termr/html/Widget.html#method-Widget-capture_mouse'><code>Widget$capture_mouse()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="child_clip"><a href='../../termr/html/Widget.html#method-Widget-child_clip'><code>Widget$child_clip()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="clear"><a href='../../termr/html/Widget.html#method-Widget-clear'><code>Widget$clear()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="computed_style"><a href='../../termr/html/Widget.html#method-Widget-computed_style'><code>Widget$computed_style()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="content_height"><a href='../../termr/html/Widget.html#method-Widget-content_height'><code>Widget$content_height()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="content_width"><a href='../../termr/html/Widget.html#method-Widget-content_width'><code>Widget$content_width()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="focus"><a href='../../termr/html/Widget.html#method-Widget-focus'><code>Widget$focus()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="format"><a href='../../termr/html/Widget.html#method-Widget-format'><code>Widget$format()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="get_state"><a href='../../termr/html/Widget.html#method-Widget-get_state'><code>Widget$get_state()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="has_class"><a href='../../termr/html/Widget.html#method-Widget-has_class'><code>Widget$has_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="invalidate"><a href='../../termr/html/Widget.html#method-Widget-invalidate'><code>Widget$invalidate()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="invalidate_paint"><a href='../../termr/html/Widget.html#method-Widget-invalidate_paint'><code>Widget$invalidate_paint()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="is_displayed"><a href='../../termr/html/Widget.html#method-Widget-is_displayed'><code>Widget$is_displayed()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="is_enabled"><a href='../../termr/html/Widget.html#method-Widget-is_enabled'><code>Widget$is_enabled()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="layout_descend"><a href='../../termr/html/Widget.html#method-Widget-layout_descend'><code>Widget$layout_descend()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="layout_differs"><a href='../../termr/html/Widget.html#method-Widget-layout_differs'><code>Widget$layout_differs()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="matches"><a href='../../termr/html/Widget.html#method-Widget-matches'><code>Widget$matches()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="mount"><a href='../../termr/html/Widget.html#method-Widget-mount'><code>Widget$mount()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="on"><a href='../../termr/html/Widget.html#method-Widget-on'><code>Widget$on()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="paint"><a href='../../termr/html/Widget.html#method-Widget-paint'><code>Widget$paint()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="post_message"><a href='../../termr/html/Widget.html#method-Widget-post_message'><code>Widget$post_message()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="print"><a href='../../termr/html/Widget.html#method-Widget-print'><code>Widget$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="pseudo_states"><a href='../../termr/html/Widget.html#method-Widget-pseudo_states'><code>Widget$pseudo_states()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="query"><a href='../../termr/html/Widget.html#method-Widget-query'><code>Widget$query()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="query_one"><a href='../../termr/html/Widget.html#method-Widget-query_one'><code>Widget$query_one()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="refresh"><a href='../../termr/html/Widget.html#method-Widget-refresh'><code>Widget$refresh()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="release_mouse"><a href='../../termr/html/Widget.html#method-Widget-release_mouse'><code>Widget$release_mouse()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="remove"><a href='../../termr/html/Widget.html#method-Widget-remove'><code>Widget$remove()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="remove_children"><a href='../../termr/html/Widget.html#method-Widget-remove_children'><code>Widget$remove_children()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="remove_class"><a href='../../termr/html/Widget.html#method-Widget-remove_class'><code>Widget$remove_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="render"><a href='../../termr/html/Widget.html#method-Widget-render'><code>Widget$render()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="render_lines"><a href='../../termr/html/Widget.html#method-Widget-render_lines'><code>Widget$render_lines()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="replace"><a href='../../termr/html/Widget.html#method-Widget-replace'><code>Widget$replace()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="run_action"><a href='../../termr/html/Widget.html#method-Widget-run_action'><code>Widget$run_action()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="run_process"><a href='../../termr/html/Widget.html#method-Widget-run_process'><code>Widget$run_process()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="run_worker"><a href='../../termr/html/Widget.html#method-Widget-run_worker'><code>Widget$run_worker()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set"><a href='../../termr/html/Widget.html#method-Widget-set'><code>Widget$set()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set_interval"><a href='../../termr/html/Widget.html#method-Widget-set_interval'><code>Widget$set_interval()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set_state"><a href='../../termr/html/Widget.html#method-Widget-set_state'><code>Widget$set_state()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set_timeout"><a href='../../termr/html/Widget.html#method-Widget-set_timeout'><code>Widget$set_timeout()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="toggle_class"><a href='../../termr/html/Widget.html#method-Widget-toggle_class'><code>Widget$toggle_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="walk"><a href='../../termr/html/Widget.html#method-Widget-walk'><code>Widget$walk()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Vertical" data-id="default_style"><a href='../../termr/html/Vertical.html#method-Vertical-default_style'><code>Vertical$default_style()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProcessView-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-ProcessView-initialize}{}}}
\subsection{\code{ProcessView$new()}}{
  Create a process view. See \code{\link[=process_view]{process_view()}}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ProcessView$new(
  command,
  args = character(),
  wd = NULL,
  env = NULL,
  timeout = NULL,
  autostart = TRUE,
  max_lines = 2000L,
  id = NULL,
  classes = NULL,
  style = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{command, args, wd, env, timeout}}{The process to run.}
      \item{\code{autostart}}{Start when mounted?}
      \item{\code{max_lines}}{Output lines kept.}
      \item{\code{id, classes, style}}{See \link{Widget}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProcessView-default_bindings"></a>}}
\if{latex}{\out{\hypertarget{method-ProcessView-default_bindings}{}}}
\subsection{\code{ProcessView$default_bindings()}}{
  Ctrl+X cancels the process, Ctrl+R restarts it.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ProcessView$default_bindings()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProcessView-on_mount"></a>}}
\if{latex}{\out{\hypertarget{method-ProcessView-on_mount}{}}}
\subsection{\code{ProcessView$on_mount()}}{
  Start the process when the widget is mounted.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ProcessView$on_mount(event)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{event}}{A \code{MountEvent}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProcessView-start"></a>}}
\if{latex}{\out{\hypertarget{method-ProcessView-start}{}}}
\subsection{\code{ProcessView$start()}}{
  Run the process (output starts afresh).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ProcessView$start()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProcessView-cancel"></a>}}
\if{latex}{\out{\hypertarget{method-ProcessView-cancel}{}}}
\subsection{\code{ProcessView$cancel()}}{
  Stop the process.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ProcessView$cancel()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProcessView-restart"></a>}}
\if{latex}{\out{\hypertarget{method-ProcessView-restart}{}}}
\subsection{\code{ProcessView$restart()}}{
  Cancel the process and run it again.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ProcessView$restart()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProcessView-action_cancel_process"></a>}}
\if{latex}{\out{\hypertarget{method-ProcessView-action_cancel_process}{}}}
\subsection{\code{ProcessView$action_cancel_process()}}{
  Cancel action.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ProcessView$action_cancel_process()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProcessView-action_restart_process"></a>}}
\if{latex}{\out{\hypertarget{method-ProcessView-action_restart_process}{}}}
\subsection{\code{ProcessView$action_restart_process()}}{
  Restart action.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ProcessView$action_restart_process()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProcessView-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProcessView-clone}{}}}
\subsection{\code{ProcessView$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ProcessView$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
