% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-workflows.R
\name{PropertyGrid}
\alias{PropertyGrid}
\title{PropertyGrid widget}
\description{
See \code{\link[=property_grid]{property_grid()}}.
}
\section{Super classes}{
\code{\link[termr:Widget]{Widget}} -> \code{\link[termr:ScrollView]{ScrollView}} -> \code{PropertyGrid}
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{max_key_width}}{Maximum key width in terminal cells.}

    \item{\code{formatter}}{Optional \verb{function(name, value)}.}
  }
  \if{html}{\out{</div>}}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{data}}{The current named data (reactive).}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-PropertyGrid-initialize}{\code{PropertyGrid$new()}}
    \item \href{#method-PropertyGrid-default_style}{\code{PropertyGrid$default_style()}}
    \item \href{#method-PropertyGrid-set_data}{\code{PropertyGrid$set_data()}}
    \item \href{#method-PropertyGrid-clone}{\code{PropertyGrid$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="add_class"><a href='../../termr/html/Widget.html#method-Widget-add_class'><code>Widget$add_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="ancestors"><a href='../../termr/html/Widget.html#method-Widget-ancestors'><code>Widget$ancestors()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="animate"><a href='../../termr/html/Widget.html#method-Widget-animate'><code>Widget$animate()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="bind"><a href='../../termr/html/Widget.html#method-Widget-bind'><code>Widget$bind()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="bind_reactive"><a href='../../termr/html/Widget.html#method-Widget-bind_reactive'><code>Widget$bind_reactive()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="bindings"><a href='../../termr/html/Widget.html#method-Widget-bindings'><code>Widget$bindings()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="blur"><a href='../../termr/html/Widget.html#method-Widget-blur'><code>Widget$blur()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="capture_mouse"><a href='../../termr/html/Widget.html#method-Widget-capture_mouse'><code>Widget$capture_mouse()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="clear"><a href='../../termr/html/Widget.html#method-Widget-clear'><code>Widget$clear()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="computed_style"><a href='../../termr/html/Widget.html#method-Widget-computed_style'><code>Widget$computed_style()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="content_height"><a href='../../termr/html/Widget.html#method-Widget-content_height'><code>Widget$content_height()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="content_width"><a href='../../termr/html/Widget.html#method-Widget-content_width'><code>Widget$content_width()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="focus"><a href='../../termr/html/Widget.html#method-Widget-focus'><code>Widget$focus()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="format"><a href='../../termr/html/Widget.html#method-Widget-format'><code>Widget$format()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="get_state"><a href='../../termr/html/Widget.html#method-Widget-get_state'><code>Widget$get_state()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="has_class"><a href='../../termr/html/Widget.html#method-Widget-has_class'><code>Widget$has_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="invalidate"><a href='../../termr/html/Widget.html#method-Widget-invalidate'><code>Widget$invalidate()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="invalidate_paint"><a href='../../termr/html/Widget.html#method-Widget-invalidate_paint'><code>Widget$invalidate_paint()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="is_displayed"><a href='../../termr/html/Widget.html#method-Widget-is_displayed'><code>Widget$is_displayed()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="is_enabled"><a href='../../termr/html/Widget.html#method-Widget-is_enabled'><code>Widget$is_enabled()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="layout_differs"><a href='../../termr/html/Widget.html#method-Widget-layout_differs'><code>Widget$layout_differs()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="matches"><a href='../../termr/html/Widget.html#method-Widget-matches'><code>Widget$matches()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="mount"><a href='../../termr/html/Widget.html#method-Widget-mount'><code>Widget$mount()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="on"><a href='../../termr/html/Widget.html#method-Widget-on'><code>Widget$on()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="post_message"><a href='../../termr/html/Widget.html#method-Widget-post_message'><code>Widget$post_message()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="print"><a href='../../termr/html/Widget.html#method-Widget-print'><code>Widget$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="pseudo_states"><a href='../../termr/html/Widget.html#method-Widget-pseudo_states'><code>Widget$pseudo_states()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="query"><a href='../../termr/html/Widget.html#method-Widget-query'><code>Widget$query()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="query_one"><a href='../../termr/html/Widget.html#method-Widget-query_one'><code>Widget$query_one()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="refresh"><a href='../../termr/html/Widget.html#method-Widget-refresh'><code>Widget$refresh()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="release_mouse"><a href='../../termr/html/Widget.html#method-Widget-release_mouse'><code>Widget$release_mouse()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="remove"><a href='../../termr/html/Widget.html#method-Widget-remove'><code>Widget$remove()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="remove_children"><a href='../../termr/html/Widget.html#method-Widget-remove_children'><code>Widget$remove_children()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="remove_class"><a href='../../termr/html/Widget.html#method-Widget-remove_class'><code>Widget$remove_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="render"><a href='../../termr/html/Widget.html#method-Widget-render'><code>Widget$render()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="render_lines"><a href='../../termr/html/Widget.html#method-Widget-render_lines'><code>Widget$render_lines()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="replace"><a href='../../termr/html/Widget.html#method-Widget-replace'><code>Widget$replace()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="run_action"><a href='../../termr/html/Widget.html#method-Widget-run_action'><code>Widget$run_action()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="run_process"><a href='../../termr/html/Widget.html#method-Widget-run_process'><code>Widget$run_process()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="run_worker"><a href='../../termr/html/Widget.html#method-Widget-run_worker'><code>Widget$run_worker()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set"><a href='../../termr/html/Widget.html#method-Widget-set'><code>Widget$set()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set_interval"><a href='../../termr/html/Widget.html#method-Widget-set_interval'><code>Widget$set_interval()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set_state"><a href='../../termr/html/Widget.html#method-Widget-set_state'><code>Widget$set_state()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set_timeout"><a href='../../termr/html/Widget.html#method-Widget-set_timeout'><code>Widget$set_timeout()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="toggle_class"><a href='../../termr/html/Widget.html#method-Widget-toggle_class'><code>Widget$toggle_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="walk"><a href='../../termr/html/Widget.html#method-Widget-walk'><code>Widget$walk()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="ScrollView" data-id="action_page_down"><a href='../../termr/html/ScrollView.html#method-ScrollView-action_page_down'><code>ScrollView$action_page_down()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="ScrollView" data-id="action_page_up"><a href='../../termr/html/ScrollView.html#method-ScrollView-action_page_up'><code>ScrollView$action_page_up()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="ScrollView" data-id="action_scroll_down"><a href='../../termr/html/ScrollView.html#method-ScrollView-action_scroll_down'><code>ScrollView$action_scroll_down()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="ScrollView" data-id="action_scroll_end"><a href='../../termr/html/ScrollView.html#method-ScrollView-action_scroll_end'><code>ScrollView$action_scroll_end()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="ScrollView" data-id="action_scroll_home"><a href='../../termr/html/ScrollView.html#method-ScrollView-action_scroll_home'><code>ScrollView$action_scroll_home()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="ScrollView" data-id="action_scroll_left"><a href='../../termr/html/ScrollView.html#method-ScrollView-action_scroll_left'><code>ScrollView$action_scroll_left()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="ScrollView" data-id="action_scroll_right"><a href='../../termr/html/ScrollView.html#method-ScrollView-action_scroll_right'><code>ScrollView$action_scroll_right()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="ScrollView" data-id="action_scroll_up"><a href='../../termr/html/ScrollView.html#method-ScrollView-action_scroll_up'><code>ScrollView$action_scroll_up()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="ScrollView" data-id="arrange_children"><a href='../../termr/html/ScrollView.html#method-ScrollView-arrange_children'><code>ScrollView$arrange_children()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="ScrollView" data-id="child_clip"><a href='../../termr/html/ScrollView.html#method-ScrollView-child_clip'><code>ScrollView$child_clip()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="ScrollView" data-id="default_bindings"><a href='../../termr/html/ScrollView.html#method-ScrollView-default_bindings'><code>ScrollView$default_bindings()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="ScrollView" data-id="layout_descend"><a href='../../termr/html/ScrollView.html#method-ScrollView-layout_descend'><code>ScrollView$layout_descend()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="ScrollView" data-id="on_mouse_scroll"><a href='../../termr/html/ScrollView.html#method-ScrollView-on_mouse_scroll'><code>ScrollView$on_mouse_scroll()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="ScrollView" data-id="paint"><a href='../../termr/html/ScrollView.html#method-ScrollView-paint'><code>ScrollView$paint()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="ScrollView" data-id="scroll_by"><a href='../../termr/html/ScrollView.html#method-ScrollView-scroll_by'><code>ScrollView$scroll_by()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="ScrollView" data-id="scroll_end"><a href='../../termr/html/ScrollView.html#method-ScrollView-scroll_end'><code>ScrollView$scroll_end()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="ScrollView" data-id="scroll_home"><a href='../../termr/html/ScrollView.html#method-ScrollView-scroll_home'><code>ScrollView$scroll_home()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="ScrollView" data-id="scroll_into_view"><a href='../../termr/html/ScrollView.html#method-ScrollView-scroll_into_view'><code>ScrollView$scroll_into_view()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="ScrollView" data-id="scroll_to"><a href='../../termr/html/ScrollView.html#method-ScrollView-scroll_to'><code>ScrollView$scroll_to()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PropertyGrid-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-PropertyGrid-initialize}{}}}
\subsection{\code{PropertyGrid$new()}}{
  Create a property grid. See \code{\link[=property_grid]{property_grid()}}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PropertyGrid$new(
  data = list(),
  formatter = NULL,
  max_key_width = 20L,
  id = NULL,
  classes = NULL,
  style = NULL,
  disabled = FALSE
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{data}}{Named list/vector or data frame.}
      \item{\code{formatter}}{Optional value formatter.}
      \item{\code{max_key_width}}{Maximum key width.}
      \item{\code{id, classes, style, disabled}}{See \code{\link[=property_grid]{property_grid()}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PropertyGrid-default_style"></a>}}
\if{latex}{\out{\hypertarget{method-PropertyGrid-default_style}{}}}
\subsection{\code{PropertyGrid$default_style()}}{
  The built-in scrollable style.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PropertyGrid$default_style()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PropertyGrid-set_data"></a>}}
\if{latex}{\out{\hypertarget{method-PropertyGrid-set_data}{}}}
\subsection{\code{PropertyGrid$set_data()}}{
  Replace all rows.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PropertyGrid$set_data(data)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{data}}{New named data.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PropertyGrid-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PropertyGrid-clone}{}}}
\subsection{\code{PropertyGrid$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PropertyGrid$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
