% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-display.R
\name{rule}
\alias{rule}
\title{Rule (separator)}
\usage{
rule(
  orientation = "horizontal",
  title = NULL,
  id = NULL,
  classes = NULL,
  style = NULL
)
}
\arguments{
\item{orientation}{\code{"horizontal"} or \code{"vertical"}.}

\item{title}{Optional title for horizontal rules.}

\item{id}{Optional identifier.}

\item{classes}{Optional classes.}

\item{style}{A \code{\link[=style]{style()}}.}
}
\value{
A \code{Rule} widget.
}
\description{
A horizontal or vertical line, optionally with a title.
}
\examples{
render_widget(rule(title = "Results"), 20, 1)
}
