% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen-buffer.R
\name{screen_buffer}
\alias{screen_buffer}
\title{Create a virtual screen buffer}
\usage{
screen_buffer(width, height)
}
\arguments{
\item{width, height}{Size in cells.}
}
\value{
A \link{ScreenBuffer}.
}
\description{
Create a virtual screen buffer
}
\examples{
buf <- screen_buffer(10, 2)
buf$put_text(2, 1, "hello", fg = "red")
buf$to_text()
}
