% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-scroll.R
\name{ScrollView}
\alias{ScrollView}
\title{ScrollView widget}
\description{
A container whose content can be larger than the widget.
See \code{\link[=scroll_view]{scroll_view()}}.
}
\section{Super class}{
\code{\link[termr:Widget]{Widget}} -> \code{ScrollView}
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{focusable}}{Scroll views can take focus so they can be scrolled
with the keyboard even without focusable children.}

    \item{\code{direction}}{\code{"vertical"}, \code{"horizontal"} or \code{"both"}.}

    \item{\code{scrollbars}}{Draw scroll bars when the content overflows?}
  }
  \if{html}{\out{</div>}}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{offset_x, offset_y}}{Scroll offsets in cells (reactive).}

    \item{\code{viewport}}{The visible content rectangle (after the last layout).}

    \item{\code{virtual_size}}{Size of the scrollable content: \code{c(width, height)}.}

    \item{\code{max_scroll}}{Largest offsets: \code{c(x, y)}.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-ScrollView-initialize}{\code{ScrollView$new()}}
    \item \href{#method-ScrollView-default_style}{\code{ScrollView$default_style()}}
    \item \href{#method-ScrollView-default_bindings}{\code{ScrollView$default_bindings()}}
    \item \href{#method-ScrollView-scroll_to}{\code{ScrollView$scroll_to()}}
    \item \href{#method-ScrollView-scroll_by}{\code{ScrollView$scroll_by()}}
    \item \href{#method-ScrollView-scroll_home}{\code{ScrollView$scroll_home()}}
    \item \href{#method-ScrollView-scroll_end}{\code{ScrollView$scroll_end()}}
    \item \href{#method-ScrollView-scroll_into_view}{\code{ScrollView$scroll_into_view()}}
    \item \href{#method-ScrollView-on_mouse_scroll}{\code{ScrollView$on_mouse_scroll()}}
    \item \href{#method-ScrollView-action_scroll_up}{\code{ScrollView$action_scroll_up()}}
    \item \href{#method-ScrollView-action_scroll_down}{\code{ScrollView$action_scroll_down()}}
    \item \href{#method-ScrollView-action_scroll_left}{\code{ScrollView$action_scroll_left()}}
    \item \href{#method-ScrollView-action_scroll_right}{\code{ScrollView$action_scroll_right()}}
    \item \href{#method-ScrollView-action_page_up}{\code{ScrollView$action_page_up()}}
    \item \href{#method-ScrollView-action_page_down}{\code{ScrollView$action_page_down()}}
    \item \href{#method-ScrollView-action_scroll_home}{\code{ScrollView$action_scroll_home()}}
    \item \href{#method-ScrollView-action_scroll_end}{\code{ScrollView$action_scroll_end()}}
    \item \href{#method-ScrollView-arrange_children}{\code{ScrollView$arrange_children()}}
    \item \href{#method-ScrollView-layout_descend}{\code{ScrollView$layout_descend()}}
    \item \href{#method-ScrollView-child_clip}{\code{ScrollView$child_clip()}}
    \item \href{#method-ScrollView-paint}{\code{ScrollView$paint()}}
    \item \href{#method-ScrollView-clone}{\code{ScrollView$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="add_class"><a href='../../termr/html/Widget.html#method-Widget-add_class'><code>Widget$add_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="ancestors"><a href='../../termr/html/Widget.html#method-Widget-ancestors'><code>Widget$ancestors()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="animate"><a href='../../termr/html/Widget.html#method-Widget-animate'><code>Widget$animate()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="bind"><a href='../../termr/html/Widget.html#method-Widget-bind'><code>Widget$bind()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="bind_reactive"><a href='../../termr/html/Widget.html#method-Widget-bind_reactive'><code>Widget$bind_reactive()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="bindings"><a href='../../termr/html/Widget.html#method-Widget-bindings'><code>Widget$bindings()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="blur"><a href='../../termr/html/Widget.html#method-Widget-blur'><code>Widget$blur()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="capture_mouse"><a href='../../termr/html/Widget.html#method-Widget-capture_mouse'><code>Widget$capture_mouse()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="clear"><a href='../../termr/html/Widget.html#method-Widget-clear'><code>Widget$clear()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="computed_style"><a href='../../termr/html/Widget.html#method-Widget-computed_style'><code>Widget$computed_style()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="content_height"><a href='../../termr/html/Widget.html#method-Widget-content_height'><code>Widget$content_height()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="content_width"><a href='../../termr/html/Widget.html#method-Widget-content_width'><code>Widget$content_width()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="focus"><a href='../../termr/html/Widget.html#method-Widget-focus'><code>Widget$focus()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="format"><a href='../../termr/html/Widget.html#method-Widget-format'><code>Widget$format()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="get_state"><a href='../../termr/html/Widget.html#method-Widget-get_state'><code>Widget$get_state()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="has_class"><a href='../../termr/html/Widget.html#method-Widget-has_class'><code>Widget$has_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="invalidate"><a href='../../termr/html/Widget.html#method-Widget-invalidate'><code>Widget$invalidate()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="invalidate_paint"><a href='../../termr/html/Widget.html#method-Widget-invalidate_paint'><code>Widget$invalidate_paint()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="is_displayed"><a href='../../termr/html/Widget.html#method-Widget-is_displayed'><code>Widget$is_displayed()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="is_enabled"><a href='../../termr/html/Widget.html#method-Widget-is_enabled'><code>Widget$is_enabled()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="layout_differs"><a href='../../termr/html/Widget.html#method-Widget-layout_differs'><code>Widget$layout_differs()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="matches"><a href='../../termr/html/Widget.html#method-Widget-matches'><code>Widget$matches()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="mount"><a href='../../termr/html/Widget.html#method-Widget-mount'><code>Widget$mount()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="on"><a href='../../termr/html/Widget.html#method-Widget-on'><code>Widget$on()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="post_message"><a href='../../termr/html/Widget.html#method-Widget-post_message'><code>Widget$post_message()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="print"><a href='../../termr/html/Widget.html#method-Widget-print'><code>Widget$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="pseudo_states"><a href='../../termr/html/Widget.html#method-Widget-pseudo_states'><code>Widget$pseudo_states()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="query"><a href='../../termr/html/Widget.html#method-Widget-query'><code>Widget$query()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="query_one"><a href='../../termr/html/Widget.html#method-Widget-query_one'><code>Widget$query_one()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="refresh"><a href='../../termr/html/Widget.html#method-Widget-refresh'><code>Widget$refresh()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="release_mouse"><a href='../../termr/html/Widget.html#method-Widget-release_mouse'><code>Widget$release_mouse()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="remove"><a href='../../termr/html/Widget.html#method-Widget-remove'><code>Widget$remove()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="remove_children"><a href='../../termr/html/Widget.html#method-Widget-remove_children'><code>Widget$remove_children()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="remove_class"><a href='../../termr/html/Widget.html#method-Widget-remove_class'><code>Widget$remove_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="render"><a href='../../termr/html/Widget.html#method-Widget-render'><code>Widget$render()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="render_lines"><a href='../../termr/html/Widget.html#method-Widget-render_lines'><code>Widget$render_lines()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="replace"><a href='../../termr/html/Widget.html#method-Widget-replace'><code>Widget$replace()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="run_action"><a href='../../termr/html/Widget.html#method-Widget-run_action'><code>Widget$run_action()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="run_process"><a href='../../termr/html/Widget.html#method-Widget-run_process'><code>Widget$run_process()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="run_worker"><a href='../../termr/html/Widget.html#method-Widget-run_worker'><code>Widget$run_worker()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set"><a href='../../termr/html/Widget.html#method-Widget-set'><code>Widget$set()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set_interval"><a href='../../termr/html/Widget.html#method-Widget-set_interval'><code>Widget$set_interval()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set_state"><a href='../../termr/html/Widget.html#method-Widget-set_state'><code>Widget$set_state()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set_timeout"><a href='../../termr/html/Widget.html#method-Widget-set_timeout'><code>Widget$set_timeout()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="toggle_class"><a href='../../termr/html/Widget.html#method-Widget-toggle_class'><code>Widget$toggle_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="walk"><a href='../../termr/html/Widget.html#method-Widget-walk'><code>Widget$walk()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScrollView-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-ScrollView-initialize}{}}}
\subsection{\code{ScrollView$new()}}{
  Create a scroll view. See \code{\link[=scroll_view]{scroll_view()}}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScrollView$new(
  ...,
  direction = "vertical",
  scrollbars = TRUE,
  focusable = TRUE,
  id = NULL,
  classes = NULL,
  style = NULL,
  disabled = FALSE
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{Child widgets.}
      \item{\code{direction}}{Scroll direction.}
      \item{\code{scrollbars}}{Show scroll bars?}
      \item{\code{focusable}}{Can the scroll view itself be focused?}
      \item{\code{id, classes, style, disabled}}{See \link{Widget}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScrollView-default_style"></a>}}
\if{latex}{\out{\hypertarget{method-ScrollView-default_style}{}}}
\subsection{\code{ScrollView$default_style()}}{
  The built-in style.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScrollView$default_style()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScrollView-default_bindings"></a>}}
\if{latex}{\out{\hypertarget{method-ScrollView-default_bindings}{}}}
\subsection{\code{ScrollView$default_bindings()}}{
  Arrow keys, Page Up/Down and Home/End scroll.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScrollView$default_bindings()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScrollView-scroll_to"></a>}}
\if{latex}{\out{\hypertarget{method-ScrollView-scroll_to}{}}}
\subsection{\code{ScrollView$scroll_to()}}{
  Scroll to an absolute position (in cells). Positions
are clamped to the scrollable range at the next layout.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScrollView$scroll_to(x = NULL, y = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{x, y}}{New offsets; \code{NULL} keeps the current one.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScrollView-scroll_by"></a>}}
\if{latex}{\out{\hypertarget{method-ScrollView-scroll_by}{}}}
\subsection{\code{ScrollView$scroll_by()}}{
  Scroll relative to the current position.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScrollView$scroll_by(dx = 0L, dy = 0L)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{dx, dy}}{Cells to scroll (negative: up/left).}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScrollView-scroll_home"></a>}}
\if{latex}{\out{\hypertarget{method-ScrollView-scroll_home}{}}}
\subsection{\code{ScrollView$scroll_home()}}{
  Scroll to the top (and left).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScrollView$scroll_home()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScrollView-scroll_end"></a>}}
\if{latex}{\out{\hypertarget{method-ScrollView-scroll_end}{}}}
\subsection{\code{ScrollView$scroll_end()}}{
  Scroll to the bottom.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScrollView$scroll_end()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScrollView-scroll_into_view"></a>}}
\if{latex}{\out{\hypertarget{method-ScrollView-scroll_into_view}{}}}
\subsection{\code{ScrollView$scroll_into_view()}}{
  Scroll so that a descendant is fully visible (as far as
possible). Uses the last layout; the app re-lays out afterwards.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScrollView$scroll_into_view(widget)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{widget}}{A descendant widget.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \code{TRUE} if the offsets changed.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScrollView-on_mouse_scroll"></a>}}
\if{latex}{\out{\hypertarget{method-ScrollView-on_mouse_scroll}{}}}
\subsection{\code{ScrollView$on_mouse_scroll()}}{
  The mouse wheel scrolls by three lines. When the view
cannot scroll further, the event bubbles to outer scroll views.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScrollView$on_mouse_scroll(event)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{event}}{A \code{MouseEvent}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScrollView-action_scroll_up"></a>}}
\if{latex}{\out{\hypertarget{method-ScrollView-action_scroll_up}{}}}
\subsection{\code{ScrollView$action_scroll_up()}}{
  Scroll one line up.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScrollView$action_scroll_up()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScrollView-action_scroll_down"></a>}}
\if{latex}{\out{\hypertarget{method-ScrollView-action_scroll_down}{}}}
\subsection{\code{ScrollView$action_scroll_down()}}{
  Scroll one line down.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScrollView$action_scroll_down()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScrollView-action_scroll_left"></a>}}
\if{latex}{\out{\hypertarget{method-ScrollView-action_scroll_left}{}}}
\subsection{\code{ScrollView$action_scroll_left()}}{
  Scroll one column left.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScrollView$action_scroll_left()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScrollView-action_scroll_right"></a>}}
\if{latex}{\out{\hypertarget{method-ScrollView-action_scroll_right}{}}}
\subsection{\code{ScrollView$action_scroll_right()}}{
  Scroll one column right.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScrollView$action_scroll_right()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScrollView-action_page_up"></a>}}
\if{latex}{\out{\hypertarget{method-ScrollView-action_page_up}{}}}
\subsection{\code{ScrollView$action_page_up()}}{
  Scroll one page up.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScrollView$action_page_up()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScrollView-action_page_down"></a>}}
\if{latex}{\out{\hypertarget{method-ScrollView-action_page_down}{}}}
\subsection{\code{ScrollView$action_page_down()}}{
  Scroll one page down.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScrollView$action_page_down()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScrollView-action_scroll_home"></a>}}
\if{latex}{\out{\hypertarget{method-ScrollView-action_scroll_home}{}}}
\subsection{\code{ScrollView$action_scroll_home()}}{
  Scroll to the top.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScrollView$action_scroll_home()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScrollView-action_scroll_end"></a>}}
\if{latex}{\out{\hypertarget{method-ScrollView-action_scroll_end}{}}}
\subsection{\code{ScrollView$action_scroll_end()}}{
  Scroll to the bottom.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScrollView$action_scroll_end()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScrollView-arrange_children"></a>}}
\if{latex}{\out{\hypertarget{method-ScrollView-arrange_children}{}}}
\subsection{\code{ScrollView$arrange_children()}}{
  Lay the children out in a virtual area, shifted by the
scroll offsets; reserves space for scroll bars when needed.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScrollView$arrange_children(children, inner, st)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{children}}{Visible children.}
      \item{\code{inner}}{Content rectangle.}
      \item{\code{st}}{Computed style.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScrollView-layout_descend"></a>}}
\if{latex}{\out{\hypertarget{method-ScrollView-layout_descend}{}}}
\subsection{\code{ScrollView$layout_descend()}}{
  Children entirely outside the viewport are not laid out
in detail (unless a complete layout is needed, e.g. to scroll a
focused widget into view).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScrollView$layout_descend(rects)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{rects}}{Child rectangles.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScrollView-child_clip"></a>}}
\if{latex}{\out{\hypertarget{method-ScrollView-child_clip}{}}}
\subsection{\code{ScrollView$child_clip()}}{
  Children are clipped to the viewport.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScrollView$child_clip(st)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{st}}{Computed style.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScrollView-paint"></a>}}
\if{latex}{\out{\hypertarget{method-ScrollView-paint}{}}}
\subsection{\code{ScrollView$paint()}}{
  Paint background, border and scroll bars.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScrollView$paint(buffer, area, st)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{buffer}}{A \link{ScreenBuffer}.}
      \item{\code{area}}{Visible part of the region.}
      \item{\code{st}}{Computed style.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScrollView-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ScrollView-clone}{}}}
\subsection{\code{ScrollView$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScrollView$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
