% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-split.R
\name{SplitHandle}
\alias{SplitHandle}
\title{Split handle}
\description{
The draggable divider of a \code{\link[=split_pane]{split_pane()}}.
}
\keyword{internal}
\section{Super class}{
\code{\link[termr:Widget]{Widget}} -> \code{SplitHandle}
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{focusable}}{The divider can be moved with the keyboard.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-SplitHandle-default_style}{\code{SplitHandle$default_style()}}
    \item \href{#method-SplitHandle-default_bindings}{\code{SplitHandle$default_bindings()}}
    \item \href{#method-SplitHandle-action_split_less}{\code{SplitHandle$action_split_less()}}
    \item \href{#method-SplitHandle-action_split_more}{\code{SplitHandle$action_split_more()}}
    \item \href{#method-SplitHandle-action_split_less_far}{\code{SplitHandle$action_split_less_far()}}
    \item \href{#method-SplitHandle-action_split_more_far}{\code{SplitHandle$action_split_more_far()}}
    \item \href{#method-SplitHandle-action_split_min}{\code{SplitHandle$action_split_min()}}
    \item \href{#method-SplitHandle-action_split_max}{\code{SplitHandle$action_split_max()}}
    \item \href{#method-SplitHandle-on_drag_move}{\code{SplitHandle$on_drag_move()}}
    \item \href{#method-SplitHandle-on_mouse_down}{\code{SplitHandle$on_mouse_down()}}
    \item \href{#method-SplitHandle-on_drag_end}{\code{SplitHandle$on_drag_end()}}
    \item \href{#method-SplitHandle-paint}{\code{SplitHandle$paint()}}
    \item \href{#method-SplitHandle-clone}{\code{SplitHandle$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="add_class"><a href='../../termr/html/Widget.html#method-Widget-add_class'><code>Widget$add_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="ancestors"><a href='../../termr/html/Widget.html#method-Widget-ancestors'><code>Widget$ancestors()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="animate"><a href='../../termr/html/Widget.html#method-Widget-animate'><code>Widget$animate()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="arrange_children"><a href='../../termr/html/Widget.html#method-Widget-arrange_children'><code>Widget$arrange_children()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="bind"><a href='../../termr/html/Widget.html#method-Widget-bind'><code>Widget$bind()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="bind_reactive"><a href='../../termr/html/Widget.html#method-Widget-bind_reactive'><code>Widget$bind_reactive()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="bindings"><a href='../../termr/html/Widget.html#method-Widget-bindings'><code>Widget$bindings()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="blur"><a href='../../termr/html/Widget.html#method-Widget-blur'><code>Widget$blur()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="capture_mouse"><a href='../../termr/html/Widget.html#method-Widget-capture_mouse'><code>Widget$capture_mouse()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="child_clip"><a href='../../termr/html/Widget.html#method-Widget-child_clip'><code>Widget$child_clip()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="clear"><a href='../../termr/html/Widget.html#method-Widget-clear'><code>Widget$clear()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="computed_style"><a href='../../termr/html/Widget.html#method-Widget-computed_style'><code>Widget$computed_style()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="content_height"><a href='../../termr/html/Widget.html#method-Widget-content_height'><code>Widget$content_height()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="content_width"><a href='../../termr/html/Widget.html#method-Widget-content_width'><code>Widget$content_width()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="focus"><a href='../../termr/html/Widget.html#method-Widget-focus'><code>Widget$focus()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="format"><a href='../../termr/html/Widget.html#method-Widget-format'><code>Widget$format()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="get_state"><a href='../../termr/html/Widget.html#method-Widget-get_state'><code>Widget$get_state()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="has_class"><a href='../../termr/html/Widget.html#method-Widget-has_class'><code>Widget$has_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="initialize"><a href='../../termr/html/Widget.html#method-Widget-initialize'><code>Widget$initialize()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="invalidate"><a href='../../termr/html/Widget.html#method-Widget-invalidate'><code>Widget$invalidate()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="invalidate_paint"><a href='../../termr/html/Widget.html#method-Widget-invalidate_paint'><code>Widget$invalidate_paint()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="is_displayed"><a href='../../termr/html/Widget.html#method-Widget-is_displayed'><code>Widget$is_displayed()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="is_enabled"><a href='../../termr/html/Widget.html#method-Widget-is_enabled'><code>Widget$is_enabled()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="layout_descend"><a href='../../termr/html/Widget.html#method-Widget-layout_descend'><code>Widget$layout_descend()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="layout_differs"><a href='../../termr/html/Widget.html#method-Widget-layout_differs'><code>Widget$layout_differs()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="matches"><a href='../../termr/html/Widget.html#method-Widget-matches'><code>Widget$matches()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="mount"><a href='../../termr/html/Widget.html#method-Widget-mount'><code>Widget$mount()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="on"><a href='../../termr/html/Widget.html#method-Widget-on'><code>Widget$on()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="post_message"><a href='../../termr/html/Widget.html#method-Widget-post_message'><code>Widget$post_message()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="print"><a href='../../termr/html/Widget.html#method-Widget-print'><code>Widget$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="pseudo_states"><a href='../../termr/html/Widget.html#method-Widget-pseudo_states'><code>Widget$pseudo_states()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="query"><a href='../../termr/html/Widget.html#method-Widget-query'><code>Widget$query()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="query_one"><a href='../../termr/html/Widget.html#method-Widget-query_one'><code>Widget$query_one()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="refresh"><a href='../../termr/html/Widget.html#method-Widget-refresh'><code>Widget$refresh()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="release_mouse"><a href='../../termr/html/Widget.html#method-Widget-release_mouse'><code>Widget$release_mouse()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="remove"><a href='../../termr/html/Widget.html#method-Widget-remove'><code>Widget$remove()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="remove_children"><a href='../../termr/html/Widget.html#method-Widget-remove_children'><code>Widget$remove_children()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="remove_class"><a href='../../termr/html/Widget.html#method-Widget-remove_class'><code>Widget$remove_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="render"><a href='../../termr/html/Widget.html#method-Widget-render'><code>Widget$render()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="render_lines"><a href='../../termr/html/Widget.html#method-Widget-render_lines'><code>Widget$render_lines()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="replace"><a href='../../termr/html/Widget.html#method-Widget-replace'><code>Widget$replace()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="run_action"><a href='../../termr/html/Widget.html#method-Widget-run_action'><code>Widget$run_action()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="run_process"><a href='../../termr/html/Widget.html#method-Widget-run_process'><code>Widget$run_process()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="run_worker"><a href='../../termr/html/Widget.html#method-Widget-run_worker'><code>Widget$run_worker()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set"><a href='../../termr/html/Widget.html#method-Widget-set'><code>Widget$set()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set_interval"><a href='../../termr/html/Widget.html#method-Widget-set_interval'><code>Widget$set_interval()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set_state"><a href='../../termr/html/Widget.html#method-Widget-set_state'><code>Widget$set_state()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set_timeout"><a href='../../termr/html/Widget.html#method-Widget-set_timeout'><code>Widget$set_timeout()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="toggle_class"><a href='../../termr/html/Widget.html#method-Widget-toggle_class'><code>Widget$toggle_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="walk"><a href='../../termr/html/Widget.html#method-Widget-walk'><code>Widget$walk()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SplitHandle-default_style"></a>}}
\if{latex}{\out{\hypertarget{method-SplitHandle-default_style}{}}}
\subsection{\code{SplitHandle$default_style()}}{
  The built-in style.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SplitHandle$default_style()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SplitHandle-default_bindings"></a>}}
\if{latex}{\out{\hypertarget{method-SplitHandle-default_bindings}{}}}
\subsection{\code{SplitHandle$default_bindings()}}{
  Arrow keys move the divider (Shift: ten cells).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SplitHandle$default_bindings()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SplitHandle-action_split_less"></a>}}
\if{latex}{\out{\hypertarget{method-SplitHandle-action_split_less}{}}}
\subsection{\code{SplitHandle$action_split_less()}}{
  Move the divider towards the first pane.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SplitHandle$action_split_less()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SplitHandle-action_split_more"></a>}}
\if{latex}{\out{\hypertarget{method-SplitHandle-action_split_more}{}}}
\subsection{\code{SplitHandle$action_split_more()}}{
  Move the divider towards the second pane.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SplitHandle$action_split_more()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SplitHandle-action_split_less_far"></a>}}
\if{latex}{\out{\hypertarget{method-SplitHandle-action_split_less_far}{}}}
\subsection{\code{SplitHandle$action_split_less_far()}}{
  Move the divider ten cells towards the first pane.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SplitHandle$action_split_less_far()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SplitHandle-action_split_more_far"></a>}}
\if{latex}{\out{\hypertarget{method-SplitHandle-action_split_more_far}{}}}
\subsection{\code{SplitHandle$action_split_more_far()}}{
  Move the divider ten cells towards the second pane.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SplitHandle$action_split_more_far()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SplitHandle-action_split_min"></a>}}
\if{latex}{\out{\hypertarget{method-SplitHandle-action_split_min}{}}}
\subsection{\code{SplitHandle$action_split_min()}}{
  Give the first pane its smallest size.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SplitHandle$action_split_min()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SplitHandle-action_split_max"></a>}}
\if{latex}{\out{\hypertarget{method-SplitHandle-action_split_max}{}}}
\subsection{\code{SplitHandle$action_split_max()}}{
  Give the first pane its largest size.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SplitHandle$action_split_max()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SplitHandle-on_drag_move"></a>}}
\if{latex}{\out{\hypertarget{method-SplitHandle-on_drag_move}{}}}
\subsection{\code{SplitHandle$on_drag_move()}}{
  Dragging the divider resizes the panes.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SplitHandle$on_drag_move(event)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{event}}{A \code{MouseEvent}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SplitHandle-on_mouse_down"></a>}}
\if{latex}{\out{\hypertarget{method-SplitHandle-on_mouse_down}{}}}
\subsection{\code{SplitHandle$on_mouse_down()}}{
  A press on the divider starts a drag.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SplitHandle$on_mouse_down(event)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{event}}{A \code{MouseEvent}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SplitHandle-on_drag_end"></a>}}
\if{latex}{\out{\hypertarget{method-SplitHandle-on_drag_end}{}}}
\subsection{\code{SplitHandle$on_drag_end()}}{
  Announce the new size when a drag ends.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SplitHandle$on_drag_end(event)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{event}}{A \code{MouseEvent}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SplitHandle-paint"></a>}}
\if{latex}{\out{\hypertarget{method-SplitHandle-paint}{}}}
\subsection{\code{SplitHandle$paint()}}{
  Draw the divider line.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SplitHandle$paint(buffer, area, st)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{buffer}}{A \link{ScreenBuffer}.}
      \item{\code{area}}{Visible part of the region.}
      \item{\code{st}}{Computed style.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SplitHandle-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SplitHandle-clone}{}}}
\subsection{\code{SplitHandle$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SplitHandle$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
