% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ansi.R
\name{strip_ansi}
\alias{strip_ansi}
\title{Remove ANSI escape sequences from a string}
\usage{
strip_ansi(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
\code{x} without CSI/OSC escape sequences.
}
\description{
Remove ANSI escape sequences from a string
}
\examples{
strip_ansi("\033[1mbold\033[0m")
}
