% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-tabs.R
\name{Tabs}
\alias{Tabs}
\title{Tabs widget}
\description{
Tabbed content. See \code{\link[=tabs]{tabs()}}.
}
\section{Super class}{
\code{\link[termr:Widget]{Widget}} -> \code{Tabs}
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{focusable}}{The tab bar can take focus (Left/Right switch tabs).}
  }
  \if{html}{\out{</div>}}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{active}}{Index of the active tab (assign to switch).}

    \item{\code{active_tab}}{The active \link{TabPane}, or \code{NULL}.}

    \item{\code{tab_count}}{Number of tabs.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-Tabs-initialize}{\code{Tabs$new()}}
    \item \href{#method-Tabs-default_style}{\code{Tabs$default_style()}}
    \item \href{#method-Tabs-default_bindings}{\code{Tabs$default_bindings()}}
    \item \href{#method-Tabs-add_tab}{\code{Tabs$add_tab()}}
    \item \href{#method-Tabs-remove_tab}{\code{Tabs$remove_tab()}}
    \item \href{#method-Tabs-activate}{\code{Tabs$activate()}}
    \item \href{#method-Tabs-action_next_tab}{\code{Tabs$action_next_tab()}}
    \item \href{#method-Tabs-action_previous_tab}{\code{Tabs$action_previous_tab()}}
    \item \href{#method-Tabs-on_mouse_down}{\code{Tabs$on_mouse_down()}}
    \item \href{#method-Tabs-arrange_children}{\code{Tabs$arrange_children()}}
    \item \href{#method-Tabs-child_clip}{\code{Tabs$child_clip()}}
    \item \href{#method-Tabs-content_width}{\code{Tabs$content_width()}}
    \item \href{#method-Tabs-content_height}{\code{Tabs$content_height()}}
    \item \href{#method-Tabs-paint}{\code{Tabs$paint()}}
    \item \href{#method-Tabs-clone}{\code{Tabs$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="add_class"><a href='../../termr/html/Widget.html#method-Widget-add_class'><code>Widget$add_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="ancestors"><a href='../../termr/html/Widget.html#method-Widget-ancestors'><code>Widget$ancestors()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="animate"><a href='../../termr/html/Widget.html#method-Widget-animate'><code>Widget$animate()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="bind"><a href='../../termr/html/Widget.html#method-Widget-bind'><code>Widget$bind()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="bind_reactive"><a href='../../termr/html/Widget.html#method-Widget-bind_reactive'><code>Widget$bind_reactive()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="bindings"><a href='../../termr/html/Widget.html#method-Widget-bindings'><code>Widget$bindings()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="blur"><a href='../../termr/html/Widget.html#method-Widget-blur'><code>Widget$blur()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="capture_mouse"><a href='../../termr/html/Widget.html#method-Widget-capture_mouse'><code>Widget$capture_mouse()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="clear"><a href='../../termr/html/Widget.html#method-Widget-clear'><code>Widget$clear()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="computed_style"><a href='../../termr/html/Widget.html#method-Widget-computed_style'><code>Widget$computed_style()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="focus"><a href='../../termr/html/Widget.html#method-Widget-focus'><code>Widget$focus()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="format"><a href='../../termr/html/Widget.html#method-Widget-format'><code>Widget$format()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="get_state"><a href='../../termr/html/Widget.html#method-Widget-get_state'><code>Widget$get_state()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="has_class"><a href='../../termr/html/Widget.html#method-Widget-has_class'><code>Widget$has_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="invalidate"><a href='../../termr/html/Widget.html#method-Widget-invalidate'><code>Widget$invalidate()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="invalidate_paint"><a href='../../termr/html/Widget.html#method-Widget-invalidate_paint'><code>Widget$invalidate_paint()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="is_displayed"><a href='../../termr/html/Widget.html#method-Widget-is_displayed'><code>Widget$is_displayed()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="is_enabled"><a href='../../termr/html/Widget.html#method-Widget-is_enabled'><code>Widget$is_enabled()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="layout_descend"><a href='../../termr/html/Widget.html#method-Widget-layout_descend'><code>Widget$layout_descend()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="layout_differs"><a href='../../termr/html/Widget.html#method-Widget-layout_differs'><code>Widget$layout_differs()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="matches"><a href='../../termr/html/Widget.html#method-Widget-matches'><code>Widget$matches()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="mount"><a href='../../termr/html/Widget.html#method-Widget-mount'><code>Widget$mount()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="on"><a href='../../termr/html/Widget.html#method-Widget-on'><code>Widget$on()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="post_message"><a href='../../termr/html/Widget.html#method-Widget-post_message'><code>Widget$post_message()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="print"><a href='../../termr/html/Widget.html#method-Widget-print'><code>Widget$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="pseudo_states"><a href='../../termr/html/Widget.html#method-Widget-pseudo_states'><code>Widget$pseudo_states()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="query"><a href='../../termr/html/Widget.html#method-Widget-query'><code>Widget$query()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="query_one"><a href='../../termr/html/Widget.html#method-Widget-query_one'><code>Widget$query_one()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="refresh"><a href='../../termr/html/Widget.html#method-Widget-refresh'><code>Widget$refresh()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="release_mouse"><a href='../../termr/html/Widget.html#method-Widget-release_mouse'><code>Widget$release_mouse()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="remove"><a href='../../termr/html/Widget.html#method-Widget-remove'><code>Widget$remove()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="remove_children"><a href='../../termr/html/Widget.html#method-Widget-remove_children'><code>Widget$remove_children()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="remove_class"><a href='../../termr/html/Widget.html#method-Widget-remove_class'><code>Widget$remove_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="render"><a href='../../termr/html/Widget.html#method-Widget-render'><code>Widget$render()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="render_lines"><a href='../../termr/html/Widget.html#method-Widget-render_lines'><code>Widget$render_lines()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="replace"><a href='../../termr/html/Widget.html#method-Widget-replace'><code>Widget$replace()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="run_action"><a href='../../termr/html/Widget.html#method-Widget-run_action'><code>Widget$run_action()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="run_process"><a href='../../termr/html/Widget.html#method-Widget-run_process'><code>Widget$run_process()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="run_worker"><a href='../../termr/html/Widget.html#method-Widget-run_worker'><code>Widget$run_worker()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set"><a href='../../termr/html/Widget.html#method-Widget-set'><code>Widget$set()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set_interval"><a href='../../termr/html/Widget.html#method-Widget-set_interval'><code>Widget$set_interval()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set_state"><a href='../../termr/html/Widget.html#method-Widget-set_state'><code>Widget$set_state()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set_timeout"><a href='../../termr/html/Widget.html#method-Widget-set_timeout'><code>Widget$set_timeout()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="toggle_class"><a href='../../termr/html/Widget.html#method-Widget-toggle_class'><code>Widget$toggle_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="walk"><a href='../../termr/html/Widget.html#method-Widget-walk'><code>Widget$walk()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tabs-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-Tabs-initialize}{}}}
\subsection{\code{Tabs$new()}}{
  Create tabs. See \code{\link[=tabs]{tabs()}}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Tabs$new(
  ...,
  active = 1L,
  id = NULL,
  classes = NULL,
  style = NULL,
  disabled = FALSE
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{\code{\link[=tab]{tab()}} panes.}
      \item{\code{active}}{Index or id of the initially active tab.}
      \item{\code{id, classes, style, disabled}}{See \link{Widget}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tabs-default_style"></a>}}
\if{latex}{\out{\hypertarget{method-Tabs-default_style}{}}}
\subsection{\code{Tabs$default_style()}}{
  The built-in style.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Tabs$default_style()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tabs-default_bindings"></a>}}
\if{latex}{\out{\hypertarget{method-Tabs-default_bindings}{}}}
\subsection{\code{Tabs$default_bindings()}}{
  Left/Right switch tabs when the bar is focused;
Ctrl+PageUp / Ctrl+PageDown switch from anywhere inside.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Tabs$default_bindings()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tabs-add_tab"></a>}}
\if{latex}{\out{\hypertarget{method-Tabs-add_tab}{}}}
\subsection{\code{Tabs$add_tab()}}{
  Add a tab at the end.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Tabs$add_tab(pane, activate = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{pane}}{A \code{\link[=tab]{tab()}} (or any widget, wrapped in a tab named after
its id).}
      \item{\code{activate}}{Make it the active tab?}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tabs-remove_tab"></a>}}
\if{latex}{\out{\hypertarget{method-Tabs-remove_tab}{}}}
\subsection{\code{Tabs$remove_tab()}}{
  Remove a tab.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Tabs$remove_tab(which)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{which}}{Index or id of the tab.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tabs-activate"></a>}}
\if{latex}{\out{\hypertarget{method-Tabs-activate}{}}}
\subsection{\code{Tabs$activate()}}{
  Show a tab.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Tabs$activate(which)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{which}}{Index or id of the tab.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tabs-action_next_tab"></a>}}
\if{latex}{\out{\hypertarget{method-Tabs-action_next_tab}{}}}
\subsection{\code{Tabs$action_next_tab()}}{
  Show the next tab.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Tabs$action_next_tab()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tabs-action_previous_tab"></a>}}
\if{latex}{\out{\hypertarget{method-Tabs-action_previous_tab}{}}}
\subsection{\code{Tabs$action_previous_tab()}}{
  Show the previous tab.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Tabs$action_previous_tab()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tabs-on_mouse_down"></a>}}
\if{latex}{\out{\hypertarget{method-Tabs-on_mouse_down}{}}}
\subsection{\code{Tabs$on_mouse_down()}}{
  Clicking a label shows its tab.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Tabs$on_mouse_down(event)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{event}}{A \code{MouseEvent}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tabs-arrange_children"></a>}}
\if{latex}{\out{\hypertarget{method-Tabs-arrange_children}{}}}
\subsection{\code{Tabs$arrange_children()}}{
  Panes go below the tab bar.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Tabs$arrange_children(children, inner, st)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{children}}{Visible children.}
      \item{\code{inner}}{Content rectangle.}
      \item{\code{st}}{Computed style.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tabs-child_clip"></a>}}
\if{latex}{\out{\hypertarget{method-Tabs-child_clip}{}}}
\subsection{\code{Tabs$child_clip()}}{
  Panes are clipped below the tab bar.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Tabs$child_clip(st)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{st}}{Computed style.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tabs-content_width"></a>}}
\if{latex}{\out{\hypertarget{method-Tabs-content_width}{}}}
\subsection{\code{Tabs$content_width()}}{
  Natural width: the widest of the bar and the panes.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Tabs$content_width()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tabs-content_height"></a>}}
\if{latex}{\out{\hypertarget{method-Tabs-content_height}{}}}
\subsection{\code{Tabs$content_height()}}{
  Natural height: bar, rule and the active pane.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Tabs$content_height(width)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{width}}{Content width.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tabs-paint"></a>}}
\if{latex}{\out{\hypertarget{method-Tabs-paint}{}}}
\subsection{\code{Tabs$paint()}}{
  Paint the tab bar.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Tabs$paint(buffer, area, st)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{buffer}}{A \link{ScreenBuffer}.}
      \item{\code{area}}{Visible part of the region.}
      \item{\code{st}}{Computed style.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tabs-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Tabs-clone}{}}}
\subsection{\code{Tabs$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Tabs$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
