% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timers.R
\name{Timer}
\alias{Timer}
\title{Timer handle}
\description{
Returned by \code{\link[=set_timeout]{set_timeout()}} and \code{\link[=set_interval]{set_interval()}}.
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{delay}}{Delay (or interval) in seconds.}

    \item{\code{repeating}}{Does the timer repeat?}

    \item{\code{due}}{Clock time of the next run.}

    \item{\code{callback}}{Function called when the timer fires.}

    \item{\code{active}}{Is the timer still scheduled?}

    \item{\code{runs}}{How many times the timer has fired.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-Timer-initialize}{\code{Timer$new()}}
    \item \href{#method-Timer-cancel}{\code{Timer$cancel()}}
    \item \href{#method-Timer-print}{\code{Timer$print()}}
    \item \href{#method-Timer-clone}{\code{Timer$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Timer-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-Timer-initialize}{}}}
\subsection{\code{Timer$new()}}{
  Create a timer (use \code{\link[=set_timeout]{set_timeout()}} instead).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Timer$new(delay, callback, repeating, now)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{delay}}{Seconds.}
      \item{\code{callback}}{Function.}
      \item{\code{repeating}}{Logical.}
      \item{\code{now}}{Current clock time.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Timer-cancel"></a>}}
\if{latex}{\out{\hypertarget{method-Timer-cancel}{}}}
\subsection{\code{Timer$cancel()}}{
  Stop the timer.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Timer$cancel()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Timer-print"></a>}}
\if{latex}{\out{\hypertarget{method-Timer-print}{}}}
\subsection{\code{Timer$print()}}{
  Print the timer.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Timer$print(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{Ignored.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Timer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Timer-clone}{}}}
\subsection{\code{Timer$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Timer$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
