% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-treeview.R
\name{TreeView}
\alias{TreeView}
\title{TreeView widget}
\description{
Shows a tree of \code{\link[=tree_node]{tree_node()}}s. See \code{\link[=tree_view]{tree_view()}}.
}
\section{Super class}{
\code{\link[termr:Widget]{Widget}} -> \code{TreeView}
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{paint_states}}{Cursor and scroll changes only repaint.}

    \item{\code{focusable}}{Trees can be focused.}

    \item{\code{show_root}}{Show the root node?}

    \item{\code{guides}}{Draw guide lines?}
  }
  \if{html}{\out{</div>}}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{root}}{The root node.}

    \item{\code{cursor_node}}{The node under the cursor, or \code{NULL}.}

    \item{\code{visible_count}}{Number of visible (expanded) lines.}

    \item{\code{offset}}{First visible line minus one.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-TreeView-initialize}{\code{TreeView$new()}}
    \item \href{#method-TreeView-default_style}{\code{TreeView$default_style()}}
    \item \href{#method-TreeView-default_bindings}{\code{TreeView$default_bindings()}}
    \item \href{#method-TreeView-tree_changed}{\code{TreeView$tree_changed()}}
    \item \href{#method-TreeView-select}{\code{TreeView$select()}}
    \item \href{#method-TreeView-action_cursor_up}{\code{TreeView$action_cursor_up()}}
    \item \href{#method-TreeView-action_cursor_down}{\code{TreeView$action_cursor_down()}}
    \item \href{#method-TreeView-action_page_up}{\code{TreeView$action_page_up()}}
    \item \href{#method-TreeView-action_page_down}{\code{TreeView$action_page_down()}}
    \item \href{#method-TreeView-action_first}{\code{TreeView$action_first()}}
    \item \href{#method-TreeView-action_last}{\code{TreeView$action_last()}}
    \item \href{#method-TreeView-action_expand_or_child}{\code{TreeView$action_expand_or_child()}}
    \item \href{#method-TreeView-action_collapse_or_parent}{\code{TreeView$action_collapse_or_parent()}}
    \item \href{#method-TreeView-action_toggle}{\code{TreeView$action_toggle()}}
    \item \href{#method-TreeView-action_activate}{\code{TreeView$action_activate()}}
    \item \href{#method-TreeView-on_mouse_down}{\code{TreeView$on_mouse_down()}}
    \item \href{#method-TreeView-on_mouse_scroll}{\code{TreeView$on_mouse_scroll()}}
    \item \href{#method-TreeView-content_width}{\code{TreeView$content_width()}}
    \item \href{#method-TreeView-content_height}{\code{TreeView$content_height()}}
    \item \href{#method-TreeView-paint}{\code{TreeView$paint()}}
    \item \href{#method-TreeView-clone}{\code{TreeView$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="add_class"><a href='../../termr/html/Widget.html#method-Widget-add_class'><code>Widget$add_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="ancestors"><a href='../../termr/html/Widget.html#method-Widget-ancestors'><code>Widget$ancestors()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="animate"><a href='../../termr/html/Widget.html#method-Widget-animate'><code>Widget$animate()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="arrange_children"><a href='../../termr/html/Widget.html#method-Widget-arrange_children'><code>Widget$arrange_children()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="bind"><a href='../../termr/html/Widget.html#method-Widget-bind'><code>Widget$bind()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="bind_reactive"><a href='../../termr/html/Widget.html#method-Widget-bind_reactive'><code>Widget$bind_reactive()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="bindings"><a href='../../termr/html/Widget.html#method-Widget-bindings'><code>Widget$bindings()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="blur"><a href='../../termr/html/Widget.html#method-Widget-blur'><code>Widget$blur()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="capture_mouse"><a href='../../termr/html/Widget.html#method-Widget-capture_mouse'><code>Widget$capture_mouse()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="child_clip"><a href='../../termr/html/Widget.html#method-Widget-child_clip'><code>Widget$child_clip()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="clear"><a href='../../termr/html/Widget.html#method-Widget-clear'><code>Widget$clear()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="computed_style"><a href='../../termr/html/Widget.html#method-Widget-computed_style'><code>Widget$computed_style()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="focus"><a href='../../termr/html/Widget.html#method-Widget-focus'><code>Widget$focus()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="format"><a href='../../termr/html/Widget.html#method-Widget-format'><code>Widget$format()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="get_state"><a href='../../termr/html/Widget.html#method-Widget-get_state'><code>Widget$get_state()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="has_class"><a href='../../termr/html/Widget.html#method-Widget-has_class'><code>Widget$has_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="invalidate"><a href='../../termr/html/Widget.html#method-Widget-invalidate'><code>Widget$invalidate()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="invalidate_paint"><a href='../../termr/html/Widget.html#method-Widget-invalidate_paint'><code>Widget$invalidate_paint()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="is_displayed"><a href='../../termr/html/Widget.html#method-Widget-is_displayed'><code>Widget$is_displayed()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="is_enabled"><a href='../../termr/html/Widget.html#method-Widget-is_enabled'><code>Widget$is_enabled()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="layout_descend"><a href='../../termr/html/Widget.html#method-Widget-layout_descend'><code>Widget$layout_descend()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="layout_differs"><a href='../../termr/html/Widget.html#method-Widget-layout_differs'><code>Widget$layout_differs()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="matches"><a href='../../termr/html/Widget.html#method-Widget-matches'><code>Widget$matches()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="mount"><a href='../../termr/html/Widget.html#method-Widget-mount'><code>Widget$mount()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="on"><a href='../../termr/html/Widget.html#method-Widget-on'><code>Widget$on()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="post_message"><a href='../../termr/html/Widget.html#method-Widget-post_message'><code>Widget$post_message()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="print"><a href='../../termr/html/Widget.html#method-Widget-print'><code>Widget$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="pseudo_states"><a href='../../termr/html/Widget.html#method-Widget-pseudo_states'><code>Widget$pseudo_states()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="query"><a href='../../termr/html/Widget.html#method-Widget-query'><code>Widget$query()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="query_one"><a href='../../termr/html/Widget.html#method-Widget-query_one'><code>Widget$query_one()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="refresh"><a href='../../termr/html/Widget.html#method-Widget-refresh'><code>Widget$refresh()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="release_mouse"><a href='../../termr/html/Widget.html#method-Widget-release_mouse'><code>Widget$release_mouse()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="remove"><a href='../../termr/html/Widget.html#method-Widget-remove'><code>Widget$remove()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="remove_children"><a href='../../termr/html/Widget.html#method-Widget-remove_children'><code>Widget$remove_children()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="remove_class"><a href='../../termr/html/Widget.html#method-Widget-remove_class'><code>Widget$remove_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="render"><a href='../../termr/html/Widget.html#method-Widget-render'><code>Widget$render()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="render_lines"><a href='../../termr/html/Widget.html#method-Widget-render_lines'><code>Widget$render_lines()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="replace"><a href='../../termr/html/Widget.html#method-Widget-replace'><code>Widget$replace()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="run_action"><a href='../../termr/html/Widget.html#method-Widget-run_action'><code>Widget$run_action()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="run_process"><a href='../../termr/html/Widget.html#method-Widget-run_process'><code>Widget$run_process()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="run_worker"><a href='../../termr/html/Widget.html#method-Widget-run_worker'><code>Widget$run_worker()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set"><a href='../../termr/html/Widget.html#method-Widget-set'><code>Widget$set()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set_interval"><a href='../../termr/html/Widget.html#method-Widget-set_interval'><code>Widget$set_interval()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set_state"><a href='../../termr/html/Widget.html#method-Widget-set_state'><code>Widget$set_state()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="set_timeout"><a href='../../termr/html/Widget.html#method-Widget-set_timeout'><code>Widget$set_timeout()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="toggle_class"><a href='../../termr/html/Widget.html#method-Widget-toggle_class'><code>Widget$toggle_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="termr" data-topic="Widget" data-id="walk"><a href='../../termr/html/Widget.html#method-Widget-walk'><code>Widget$walk()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeView-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-TreeView-initialize}{}}}
\subsection{\code{TreeView$new()}}{
  Create a tree view. See \code{\link[=tree_view]{tree_view()}}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TreeView$new(
  root,
  show_root = TRUE,
  guides = TRUE,
  id = NULL,
  classes = NULL,
  style = NULL,
  disabled = FALSE
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{root}}{The root \link{TreeNode}.}
      \item{\code{show_root, guides}}{See fields.}
      \item{\code{id, classes, style, disabled}}{See \link{Widget}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeView-default_style"></a>}}
\if{latex}{\out{\hypertarget{method-TreeView-default_style}{}}}
\subsection{\code{TreeView$default_style()}}{
  The built-in style.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TreeView$default_style()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeView-default_bindings"></a>}}
\if{latex}{\out{\hypertarget{method-TreeView-default_bindings}{}}}
\subsection{\code{TreeView$default_bindings()}}{
  Navigation keys.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TreeView$default_bindings()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeView-tree_changed"></a>}}
\if{latex}{\out{\hypertarget{method-TreeView-tree_changed}{}}}
\subsection{\code{TreeView$tree_changed()}}{
  Called by nodes when the tree changes.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TreeView$tree_changed()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeView-select"></a>}}
\if{latex}{\out{\hypertarget{method-TreeView-select}{}}}
\subsection{\code{TreeView$select()}}{
  Move the cursor to a node (expanding its ancestors).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TreeView$select(node)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{node}}{A node of this tree.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeView-action_cursor_up"></a>}}
\if{latex}{\out{\hypertarget{method-TreeView-action_cursor_up}{}}}
\subsection{\code{TreeView$action_cursor_up()}}{
  Move up.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TreeView$action_cursor_up()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeView-action_cursor_down"></a>}}
\if{latex}{\out{\hypertarget{method-TreeView-action_cursor_down}{}}}
\subsection{\code{TreeView$action_cursor_down()}}{
  Move down.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TreeView$action_cursor_down()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeView-action_page_up"></a>}}
\if{latex}{\out{\hypertarget{method-TreeView-action_page_up}{}}}
\subsection{\code{TreeView$action_page_up()}}{
  Move a page up.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TreeView$action_page_up()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeView-action_page_down"></a>}}
\if{latex}{\out{\hypertarget{method-TreeView-action_page_down}{}}}
\subsection{\code{TreeView$action_page_down()}}{
  Move a page down.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TreeView$action_page_down()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeView-action_first"></a>}}
\if{latex}{\out{\hypertarget{method-TreeView-action_first}{}}}
\subsection{\code{TreeView$action_first()}}{
  Go to the first node.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TreeView$action_first()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeView-action_last"></a>}}
\if{latex}{\out{\hypertarget{method-TreeView-action_last}{}}}
\subsection{\code{TreeView$action_last()}}{
  Go to the last visible node.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TreeView$action_last()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeView-action_expand_or_child"></a>}}
\if{latex}{\out{\hypertarget{method-TreeView-action_expand_or_child}{}}}
\subsection{\code{TreeView$action_expand_or_child()}}{
  Expand, or go to the first child.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TreeView$action_expand_or_child()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeView-action_collapse_or_parent"></a>}}
\if{latex}{\out{\hypertarget{method-TreeView-action_collapse_or_parent}{}}}
\subsection{\code{TreeView$action_collapse_or_parent()}}{
  Collapse, or go to the parent.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TreeView$action_collapse_or_parent()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeView-action_toggle"></a>}}
\if{latex}{\out{\hypertarget{method-TreeView-action_toggle}{}}}
\subsection{\code{TreeView$action_toggle()}}{
  Toggle the cursor node.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TreeView$action_toggle()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeView-action_activate"></a>}}
\if{latex}{\out{\hypertarget{method-TreeView-action_activate}{}}}
\subsection{\code{TreeView$action_activate()}}{
  Toggle and send \code{"tree.node_activated"}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TreeView$action_activate()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeView-on_mouse_down"></a>}}
\if{latex}{\out{\hypertarget{method-TreeView-on_mouse_down}{}}}
\subsection{\code{TreeView$on_mouse_down()}}{
  Clicking selects; clicking the arrow toggles.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TreeView$on_mouse_down(event)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{event}}{A \code{MouseEvent}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeView-on_mouse_scroll"></a>}}
\if{latex}{\out{\hypertarget{method-TreeView-on_mouse_scroll}{}}}
\subsection{\code{TreeView$on_mouse_scroll()}}{
  The wheel scrolls three lines.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TreeView$on_mouse_scroll(event)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{event}}{A \code{MouseEvent}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeView-content_width"></a>}}
\if{latex}{\out{\hypertarget{method-TreeView-content_width}{}}}
\subsection{\code{TreeView$content_width()}}{
  Natural width: the widest visible line.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TreeView$content_width()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeView-content_height"></a>}}
\if{latex}{\out{\hypertarget{method-TreeView-content_height}{}}}
\subsection{\code{TreeView$content_height()}}{
  Natural height: visible lines (at most 1000).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TreeView$content_height(width)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{width}}{Content width.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeView-paint"></a>}}
\if{latex}{\out{\hypertarget{method-TreeView-paint}{}}}
\subsection{\code{TreeView$paint()}}{
  Draw the visible lines.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TreeView$paint(buffer, area, st)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{buffer}}{A \link{ScreenBuffer}.}
      \item{\code{area}}{Visible part of the region.}
      \item{\code{st}}{Computed style.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeView-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TreeView-clone}{}}}
\subsection{\code{TreeView$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TreeView$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
